==> pg-cultivation-score-fy18/01 Initial exploration.Rmd <==
---
title: "01 Initial exploration"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Goal

Understand how the PG cultivation score, based on based on [Ben Porter's Donor Cultivation Checklist](https://www.case.org/currents/x74757), is related to future giving behavior at various capacity rating levels.

# Setup

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(readxl)
```

# Import data

The PG cultivation score was calculated from [this view](https://github.com/phively/nu-plsql/blob/243eae460f8b34d8fcc53054717976e5ac68a8c6/views/v_pg_checklist.sql), and the data file was generated by [this SQL code](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/code/pg-cultivation-score-data.sql).

```{r}
pool <- readxl::read_xlsx(
  path = 'data/2018-07-26 PG scores for all active prospects.xlsx'
  , sheet = 'With'
) %>% mutate(
  # Create factors
  QUAL_LEVEL = factor(QUAL_LEVEL)
  , PREF_ADDR = factor(PREF_ADDR)
  , PROSPECT_MGR = factor(PROSPECT_MGR)
  , MULTI_OR_SINGLE_INTEREST = factor(MULTI_OR_SINGLE_INTEREST)
  , MG_ID_MODEL_DESC = MG_ID_MODEL_DESC %>% str_remove('Major Gifts Identification ') %>% factor()
  , MG_PR_MODEL_DESC = MG_PR_MODEL_DESC %>% str_remove('Major Gifts Prioritization ') %>% factor()
  # Create booleans
  , PG_PROSPECT_FLAG = !is.na(PG_PROSPECT_FLAG)
  # Create numeric
  , MG_ID_MODEL_SCORE = as.numeric(MG_ID_MODEL_SCORE)
  , MG_ID_MODEL_YEAR = as.numeric(MG_ID_MODEL_YEAR)
  , MG_PR_MODEL_SCORE = as.numeric(MG_PR_MODEL_SCORE)
  , MG_PR_MODEL_YEAR = as.numeric(MG_PR_MODEL_YEAR)
)

# Row count
nrow(pool)
```

# Numeric summary statistics

```{r numeric_summary_functions}
# Function to compute percentiles from the empirical CDF
get_percentiles <- function(
      lbound     # smallest observation for which to compute percentile
    , ubound     # largest observation for which to compute percentile
    , group1     # First comparison group
    , group2 = 1 # Optional second comparison group
    , group3 = 1 # Optional third comparison group
  ) {
      data.frame(
        x = lbound:ubound
        , g1 = ecdf(group1)(lbound:ubound) %>% scales::percent()
        , g2 = ecdf(group2)(lbound:ubound) %>% scales::percent()
        , g3 = ecdf(group3)(lbound:ubound) %>% scales::percent()
      ) %>%
      return()
    }
```

## Cultivation score

Begin with the distribution of cultivation scores.

```{r}
lower_bound <- min(pool$CULTIVATION_SCORE)
upper_bound <- max(pool$CULTIVATION_SCORE)
summary(pool$CULTIVATION_SCORE)
```

```{r}
pool %>%
  ggplot(aes(x = CULTIVATION_SCORE)) +
  geom_histogram(binwidth = 1, alpha = .5) +
  scale_x_continuous(breaks = seq(lower_bound, upper_bound, by = 2))
```

The PG cultivation score ranges from 0 to `r upper_bound %>% I()`, with most of the data between `r quantile(pool$CULTIVATION_SCORE, .2) %>% I()` and `r quantile(pool$CULTIVATION_SCORE, .8) %>% I()`. 0 implies a totally unengaged non-alumni prospect.

Here's a look at principal gift prospects:

```{r}
pool %>%
  ggplot(aes(x = CULTIVATION_SCORE, fill = PG_PROSPECT_FLAG)) +
  geom_histogram(binwidth = 1, alpha = .5, aes(y = ..density..)) +
  scale_x_continuous(breaks = seq(lower_bound, upper_bound, by = 2)) +
  facet_grid(PG_PROSPECT_FLAG ~ .)
```

```{r, rows.print = 100}
get_percentiles(
  lower_bound
  , upper_bound
  , pool %>% select(CULTIVATION_SCORE) %>% unlist() # All prospects
  , pool %>% filter(!PG_PROSPECT_FLAG) %>% select(CULTIVATION_SCORE) %>% unlist() # Non-PG only
  , pool %>% filter(PG_PROSPECT_FLAG) %>% select(CULTIVATION_SCORE) %>% unlist() # PG only
) %>% select(score = x, overall_percentile = g1, non_PG_percentile = g2, PG_percentile = g3)
```

PG prospects have much more density in the right tail -- I wouldn't read too much into it, though, because the PG checklist items were based on historical engagement patterns of these donors.

## Affinity scores

The affinity score models general entity engagement, including student information, other relationships, and behavior.

```{r}
summary(pool$AFFINITY_SCORE)
```

There are very few NAs, which follows given that the score is recomputed periodically.

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE)) %>%
  ggplot(aes(x = AFFINITY_SCORE)) +
  geom_histogram(binwidth = 1, alpha = .5) +
  scale_x_continuous(breaks = 0:10, minor_breaks = NULL) +
  geom_vline(xintercept = mean(pool$AFFINITY_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed')
```

Most people have an affinity score at or below `r quantile(pool$AFFINITY_SCORE, .75, na.rm = TRUE) %>% I()`

Comparing affinity score and cultivation score:

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE)) %>%
  ggplot(aes(x = CULTIVATION_SCORE, y = AFFINITY_SCORE, group = factor(CULTIVATION_SCORE))) +
  geom_boxplot()
```

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE)) %>%
  group_by(CULTIVATION_SCORE, AFFINITY_SCORE) %>% summarise(n = length(ID_NUMBER)) %>%
  ggplot(aes(x = CULTIVATION_SCORE, y = AFFINITY_SCORE, fill = n)) +
  geom_tile() +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = seq(0, 12, by = 2)) +
  scale_y_continuous(breaks = seq(0, 10, by = 2)) +
  coord_equal()
```

High affinity scores are much more frequent than high cultivation scores, so the two appear to be getting at different information.

## Major gift identification scores

Let's look at similar metrics for the major gift identification score.

```{r}
summary(pool$MG_ID_MODEL_SCORE)
```

Right off the bat, we see `r pool$MG_ID_MODEL_SCORE %>% is.na() %>% sum() %>% I()` out of the `r nrow(pool) %>% I()` prospects in the file do not have this score.

```{r}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_ID_MODEL_SCORE)) +
  geom_density() +
  geom_rug(color = 'gray') +
  geom_vline(xintercept = mean(pool$MG_ID_MODEL_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed')
```

```{r}
pool %>% select(MG_ID_MODEL_SCORE) %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>% unlist() %>%
  quantile(probs = c(.25, .5, .75, .8, .85, .9, .95, .99))
```

Most modeled prospects have a model score < 0. Descriptive labels were also assigned to certain ranges of scores.

**Top tier:**

```{r}
pool %>% select(MG_ID_MODEL_SCORE, MG_ID_MODEL_DESC) %>% filter(MG_ID_MODEL_DESC == 'Top Tier') %>%
  unlist() %>% summary()
```

**Middle tier:**

```{r}
pool %>% select(MG_ID_MODEL_SCORE, MG_ID_MODEL_DESC) %>% filter(MG_ID_MODEL_DESC == 'Middle Tier') %>%
  unlist() %>% summary()
```

**Bottom tier:**

```{r}
pool %>% select(MG_ID_MODEL_SCORE, MG_ID_MODEL_DESC) %>% filter(MG_ID_MODEL_DESC == 'Bottom Tier') %>%
  unlist() %>% summary()
```

Now this is interesting -- there is quite a bit of overlap.

```{r}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_ID_MODEL_SCORE, color = MG_ID_MODEL_DESC)) +
  geom_density(alpha = .5) +
  geom_rug() +
  geom_vline(xintercept = mean(pool$MG_ID_MODEL_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed')
```

This looks a bit like a mixture model -- perhaps the model description label is the most likely class based on model score and some other input.

*Update.* I just remembered that the data file pulls in the score associated with the most recent year. Segmenting by year the pattern becomes clear:

```{r}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_ID_MODEL_SCORE, color = MG_ID_MODEL_DESC)) +
  geom_density(alpha = .5) +
  geom_rug() +
  facet_grid(MG_ID_MODEL_YEAR ~ .) +
  geom_vline(xintercept = mean(pool$MG_ID_MODEL_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed')
```

Within each year there is no overlap. High/middle/low tier must have been recomputed each year as a function of the overall distribution of scores (e.g. top 5% of scores).

We can also look at cultivation score by ID model label.

```{r}
pool %>%filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  ggplot(aes(x = CULTIVATION_SCORE, fill = MG_ID_MODEL_DESC)) +
  geom_histogram(binwidth = 1, alpha = .5, aes(y = ..density..)) +
  scale_x_continuous(breaks = seq(lower_bound, upper_bound, by = 2)) +
  facet_grid(MG_ID_MODEL_DESC ~ .)
```

```{r, rows.print = 100}
get_percentiles(
  lower_bound
  , upper_bound
  , pool %>% filter(MG_ID_MODEL_DESC == 'Bottom Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
  , pool %>% filter(MG_ID_MODEL_DESC == 'Middle Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
  , pool %>% filter(MG_ID_MODEL_DESC == 'Top Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
) %>% select(score = x, bottom_tier = g1, middle_tier = g2, top_tier = g3)
```

Very interesting -- it looks like people who have been more heavily engaged have lower cultivation scores, perhaps because they're no longer discovery prospects?

## Major gift prioritization scores

Repeat the previous analysis with the major gift prioritization scores.

```{r}
summary(pool$MG_PR_MODEL_SCORE)
```

A much higher proportion of prospects have a prioritization score than identification score. Besides the negative entries, this looks a bit like a probability model.

```{r}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_PR_MODEL_SCORE)) +
  geom_density() +
  geom_rug(color = 'gray') +
  geom_vline(xintercept = mean(pool$MG_PR_MODEL_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed')
```

Very strongly right-skewed. Here's the same plot with the scale adjusted.

```{r}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_PR_MODEL_SCORE)) +
  geom_density() +
  geom_rug(color = 'gray') +
  geom_vline(xintercept = mean(pool$MG_PR_MODEL_SCORE, na.rm = TRUE), color = 'blue', linetype = 'dashed') +
  scale_y_sqrt(breaks = 2^(0:20), minor_breaks = NULL)
```

Examining the scores by descriptive prioritization model label:

```{r}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_PR_MODEL_SCORE, color = MG_PR_MODEL_DESC)) +
  geom_density(alpha = .5) +
  geom_rug() +
  scale_y_sqrt(breaks = 2^(0:20), minor_breaks = NULL)
```

As before, there's overlap between the groups, but the long tail is entirely from the top tier group.

*Update.* The overlap again disappears when looking at distributions by year:

```{r}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  ggplot(aes(x = MG_PR_MODEL_SCORE, color = MG_PR_MODEL_DESC)) +
  geom_density(alpha = .5) +
  geom_rug() +
  facet_grid(MG_PR_MODEL_YEAR ~ .) +
  scale_y_sqrt(breaks = 2^(0:20), minor_breaks = NULL)
```

Finally, look at the distributions of cultivation scores.

```{r}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  ggplot(aes(x = CULTIVATION_SCORE, fill = MG_PR_MODEL_DESC)) +
  geom_histogram(binwidth = 1, alpha = .5, aes(y = ..density..)) +
  scale_x_continuous(breaks = seq(lower_bound, upper_bound, by = 2)) +
  facet_grid(MG_PR_MODEL_DESC ~ .)
```

```{r}
get_percentiles(
  lower_bound
  , upper_bound
  , pool %>% filter(MG_PR_MODEL_DESC == 'Bottom Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
  , pool %>% filter(MG_PR_MODEL_DESC == 'Middle Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
  , pool %>% filter(MG_PR_MODEL_DESC == 'Top Tier') %>% select(CULTIVATION_SCORE) %>% unlist()
) %>% select(score = x, bottom_tier = g1, middle_tier = g2, top_tier = g3)
```

As expected, the top tier prospects have much higher cultivation scores than the other groups.

# Relationship to giving

How are these measures related to campaign giving, or largest gift?

## Donor campaign scatterplots

First consider people who have given to the Campaign.

Below, the red lines are linear fits and the black lines are GAM fits.

```{r}
# Generic function to generate scatterplots of interest
scatterplotter <- function(data, x, y, color, trans) {
  data %>%
    ggplot(aes_string(x = x, y = y, color = color)) +
    geom_point(alpha = .5) +
    geom_smooth(color = 'black') +
    geom_smooth(method = 'lm', color = 'red') +
    scale_y_continuous(trans = trans, breaks = c(0, 10^(0:12)), labels = scales::dollar)
}
```

```{r}
pool %>% filter(!is.na(CULTIVATION_SCORE) & CAMPAIGN_NEWGIFT_CMIT_CREDIT > 0) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

At higher cultivation scores, the linear trend is about an order of magnitude below the group means.

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE) & CAMPAIGN_NEWGIFT_CMIT_CREDIT > 0) %>%
  scatterplotter(x = 'AFFINITY_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

Compared to cultivation score, the GAM fit at high affinity scores is worse -- 2 orders of magnitude off, though there are only 4 data points, as seen above.

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE) & CAMPAIGN_NEWGIFT_CMIT_CREDIT > 0) %>%
  scatterplotter(x = 'MG_ID_MODEL_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_ID_MODEL_DESC', trans = 'log10') +
  facet_grid(MG_ID_MODEL_YEAR ~ .)
```

There's no apparent relationship between MG identification model and Campaign giving.

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE) & CAMPAIGN_NEWGIFT_CMIT_CREDIT > 0) %>%
  scatterplotter(x = 'MG_PR_MODEL_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  facet_grid(MG_PR_MODEL_YEAR ~ .)
```

Looks like prioritization model scores above around .4 are quite nicely associated with campaign giving amounts, which follows, as I believe past giving is one of the inputs.

## All prospect campaign scatterplots

Repeat the above plots with all prospects, not just donors to the campaign.

```{r}
# Log10 transformation after adding 1
log10plus1_trans <- function(x) {
  scales::trans_new(
    'log10plus1'
    , transform = function(x) log10(x + 1)
    , inverse = function(x) {10^x - 1}
  )
}
```

```{r}
pool %>% filter(!is.na(CULTIVATION_SCORE)) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

Decent in the 1-7 range of scores. It's quite common to see linear giving models fall apart at the high/low end of the range.

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE)) %>%
  scatterplotter(x = 'AFFINITY_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

This looks a bit all over the place.

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  scatterplotter(x = 'MG_ID_MODEL_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_ID_MODEL_DESC', trans = 'log10plus1') +
  facet_grid(MG_ID_MODEL_YEAR ~ .)
```

Still no apparent correlation.

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  scatterplotter(x = 'MG_PR_MODEL_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  facet_grid(MG_PR_MODEL_YEAR ~ .)
```

This has the same issues as the previous plot. Needless to say, extrapolating out to infinity is a bad idea - some sort of localized fit is appropriate.

## Donor largest gift scatterplots

Repeat the above with largest gift on the y axis.

```{r}
pool %>% filter(!is.na(CULTIVATION_SCORE) & LARGEST_GIFT_OR_PAYMENT > 0) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE) & LARGEST_GIFT_OR_PAYMENT > 0) %>%
  scatterplotter(x = 'AFFINITY_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE) & LARGEST_GIFT_OR_PAYMENT > 0) %>%
  scatterplotter(x = 'MG_ID_MODEL_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_ID_MODEL_DESC', trans = 'log10') +
  facet_grid(MG_ID_MODEL_YEAR ~ .)
```

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE) & LARGEST_GIFT_OR_PAYMENT > 0) %>%
  scatterplotter(x = 'MG_PR_MODEL_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10') +
  facet_grid(MG_PR_MODEL_YEAR ~ .)
```

Largest gift or payment does not correlate as well with affinity score as Campaign giving. One possible avenue of exploration would be to look only at largest gift within the campaign counting period, but this may not be that interesting -- even with the full data, the conclusions are still the same as above.

## All prospect largest gift scatterplots

Repeat the above but leaving in the non-donors.

```{r}
pool %>% filter(!is.na(CULTIVATION_SCORE)) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

This actually looks very nice -- adding nondonors into the mix results in a nearly linear trend from low to high.

```{r}
pool %>% filter(!is.na(AFFINITY_SCORE)) %>%
  scatterplotter(x = 'AFFINITY_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL)
```

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_ID_MODEL_SCORE)) %>%
  scatterplotter(x = 'MG_ID_MODEL_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_ID_MODEL_DESC', trans = 'log10plus1') +
  facet_grid(MG_ID_MODEL_YEAR ~ .)
```

```{r, fig.height = 6}
pool %>% filter(!is.na(MG_PR_MODEL_SCORE)) %>%
  scatterplotter(x = 'MG_PR_MODEL_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', trans = 'log10plus1') +
  facet_grid(MG_PR_MODEL_YEAR ~ .)
```

None of the other plots appear noteworthy.
