==> pg-cultivation-score-fy18/02 Cultivation score weights.Rmd <==
---
title: "02 Cultivation score weights"
output:
  html_notebook:
    code_folding: hide
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Goal

Determine the extent to which each of the data elements related to the PG cultivation score (see [Ben Porter's Donor Cultivation Checklist](https://www.case.org/currents/x74757)) have impacted giving.

# Setup and datafile

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(gridExtra)
library(e1071)
library(wranglR)
library(foreach)
library(splines)

# Functions adapted from previous analysis steps
source('code/functions.R')
```

The data file is generated with [this code](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/code/generate-data.R), adapted from the import data step of [01 Initial exploration.Rmd
](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/01%20Initial%20exploration.Rmd).

```{r}
filepath <- 'data/2018-07-26 PG scores for all active prospects.xlsx'
source('code/generate-data.R')
```

# Background {#background}

The following 12 items were able to be extracted from the database:

  * Alum or spouse of alum
  * Age
  * Chicago home address
  * Visited by prospect manager in last 2 years
  * 5+ total visits
  * Deep engagement (multiple family degrees, parent, season tickets, etc.)
  * High-level annual giving (\$25K+)
  * High-level advisory board participation
  * Previous major gift (\$250K+)
  * Meeting with president
  * Open proposal
  * Consistent donor (10+ years of giving, including 1+ of last 3)

Each counts as one point toward the cultivation score, which ranges from 0 to `r max(pool$CULTIVATION_SCORE) %>% I()`.

The context for this analysis is to look at how giving is related to the cultivation score. Consider the following two plots (replicated from [01 Initial exploration.Rmd
](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/01%20Initial%20exploration.Rmd)):

```{r, echo = FALSE, message = FALSE}
pool %>% filter(!is.na(CULTIVATION_SCORE) & CAMPAIGN_NEWGIFT_CMIT_CREDIT > 0) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT'
    , color = 'MG_PR_MODEL_DESC', ytrans = 'log10', ylabels = scales::dollar) +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL) +
  labs(title = bquote('Cultivation score versus' ~ log[10] ~ 'campaign giving')
    , x = 'PG cultivation score', y = 'Campaign giving credit', color = 'MG prioritization model')
```

```{r, echo = FALSE, message = FALSE}
pool %>% filter(!is.na(CULTIVATION_SCORE)) %>%
  scatterplotter(x = 'CULTIVATION_SCORE', y = 'LARGEST_GIFT_OR_PAYMENT'
    , color = 'MG_PR_MODEL_DESC', ytrans = 'log10plus1', ylabels = scales::dollar) +
  scale_x_continuous(breaks = 0:18, minor_breaks = NULL) +
  labs(title = bquote('Cultivation score versus' ~ log[10] ~ 'largest gift')
       , x = 'PG cultivation score', y = 'Largest gift or pledge payment', color = 'MG prioritization model')
```

There is a nearly linear relationship between PG cultivation score and both campaign giving and an entity's largest gift or pledge payment. Fitting some sort of linear model will enhance our understanding of the various checklist items by estimating their relative impact.

# Data exploration

## Indicators

Look at how each of the factors are distributed.

```{r, rows.print = 100}
pool %>% select(
  ACTIVE_PROPOSALS, AGE, PM_VISIT_LAST_2_YRS, VISITS_5PLUS, AF_25K_GIFT, GAVE_IN_LAST_3_YRS
  , MG_250K_PLUS, PRESIDENT_VISIT, TRUSTEE_OR_ADVISORY_BOARD, Alumnus, DEEP_ENGAGEMENT
  , CHICAGO_HOME
) %>%
  gather('Variable', 'x', 1:12) %>%
  group_by(Variable) %>%
  summarise(pct.yes = mean(x), yes = sum(x), no = nrow(pool) - sum(x), n = nrow(pool)) %>%
  arrange(desc(pct.yes)) %>%
  mutate(pct.yes = scales::percent(pct.yes))
```

As expected, alumni status is far and away the leader. I'm a bit surprised that trustee and advisory board is as high as it is. MG and president visit are the least common factors. However, they still have hundreds of observations each so I don't see a reason to drop either.

## Underlying variables

Exploration of the underlying variables for the indicators (that were straightforward to compute). The blue lines indicate the mean, and the purple ones the median.

```{r}
# Function to produce a summary table, with higher order central moments
summary_moments <- function(x, name = NULL) {
  # Requires e1071 for the skewness and kurtosis functions
  suppressPackageStartupMessages(
    if (!require(e1071)) {
      stop('Requires installation of package e1071')
    }
  )
  # If no name passed use the name of x
  if (is.null(name)) {name <- quote(x) %>% deparse()}
  # Data frame to be returned
  data.frame(
    name = name
    , n = na.omit(x) %>% length()
    , min = min(x, na.rm = TRUE)
    , median = median(x, na.rm = TRUE)
    , mean = mean(x, na.rm = TRUE)
    , max = max(x, na.rm = TRUE)
    , sd = sd(x, na.rm = TRUE)
    , skewness = e1071::skewness(x, na.rm = TRUE)
    , kurtosis = e1071::kurtosis(x, na.rm = TRUE)
    , NAs = is.na(x) %>% sum()
  ) %>% return()
}

# Generic function to create histograms
histogrammer <- function(data, x, binwidth = 1, bins = NULL, color = NULL
                         , trans = 'identity', xbreak = waiver()) {
  vec <- paste0('data$', eval(x)) %>% parse(text = .) %>% eval()
  data %>%
    ggplot(aes_string(x = x, color = color)) +
    geom_histogram(binwidth = binwidth, bins = bins, alpha = .5) +
    geom_density(aes(y = ..count..), alpha = .5) +
    geom_vline(xintercept = mean(vec, na.rm = TRUE), color = 'blue', linetype = 'dashed') +
    geom_vline(xintercept = median(vec, na.rm = TRUE), color = 'purple', linetype = 'dotted') +
    scale_x_continuous(trans = trans, breaks = xbreak)
}
```

```{r}
pool %>% filter(!is.na(NUMERIC_AGE)) %>%
  histogrammer(x = 'NUMERIC_AGE', xbreak = seq(0, 200, by = 10)) +
  labs(title = 'Age')
```

```{r}
summary_moments(pool$NUMERIC_AGE, 'Age')
```

The age distribution is moderately right-skewed but looks fine. For a quick fix the NAs could be imputed as the group mean.

```{r}
pool %>%
  histogrammer(x = 'VISIT_COUNT') +
  labs(title = 'Visit count')
```

```{r}
summary_moments(pool$VISIT_COUNT, 'Visits')
```

It'd be sensible to transform the x axis or otherwise get rid of those outliers.

```{r}
pool %>%
  histogrammer(x = 'VISIT_COUNT', trans = 'sqrt', binwidth = NULL, bins = 200,
               xbreak = c(seq(0, 10, by = 2), seq(10, 100, by = 10), seq(100, 200, by = 20))) +
  labs(title = 'Sqrt visit, log10 count') +
  scale_y_continuous(trans = 'log10plus1', breaks = 10^(0:20))
```

Nearly a linear decrease in visit count on a log/sqrt scale - though I'm not sure how to interpret this.

```{r}
pool %>%
  histogrammer(x = 'YEARS_OF_GIVING') +
  labs(title = 'Years of giving')
```

```{r}
summary_moments(pool$YEARS_OF_GIVING, 'Years of giving')
```

This looks fine. There are more loyal donors than I would've thought.

```{r}
pool %>%
  histogrammer(x = 'YEARS_OF_GIVING_LAST_3') +
  labs(title = 'Years of giving last 3')
```

```{r}
summary_moments(pool$YEARS_OF_GIVING_LAST_3, 'Years of giving out of last 3')
```

Mostly nondonors, unsurprisingly.

```{r}
pool %>%
  histogrammer(x = 'MG_250K_COUNT') +
  labs(title = 'Count of $250K+ major gifts') +
  scale_y_continuous(trans = 'log10plus1', breaks = 10^(0:5), labels = 10^(0:5))
```

```{r}
summary_moments(pool$MG_250K_COUNT, '$250K+ major gifts')
```

```{r, rows.print = 100}
pool %>% group_by(MG_250K_COUNT) %>% mutate(n = 1) %>%
  summarise(
    total = sum(n)
    , proportion = {sum(n) / nrow(pool)} %>% scales::percent()
  )
```

Extremely few people make a single \$250K+ gift, much less multiple ones.

```{r}
pool %>%
  histogrammer(x = 'SEASON_TICKET_YEARS', xbreak = seq(0, 20, by = 2)) +
  labs(title = 'Years holding season tickets') +
  scale_y_continuous(trans = 'log10plus1', breaks = 10^(0:20))
```

```{r}
summary_moments(pool$SEASON_TICKET_YEARS, 'Years holding season tickets')
```

That jump at 10 years is odd. Perhaps season tickets have only been consistently tracked for about 10 years?

## Underlying variable scatterplots

```{r}
pool %>% filter(!is.na(NUMERIC_AGE)) %>%
  scatterplotter(x = 'NUMERIC_AGE', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(NUMERIC_AGE)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Age versus campaign giving')
```

```{r}
pool %>% filter(!is.na(NUMERIC_AGE)) %>%
  scatterplotter(x = 'NUMERIC_AGE', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(NUMERIC_AGE)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Age versus largest gift')
```

As usual, age is positively associated with giving. The outlier `r pool$NUMERIC_AGE %>% max(na.rm = TRUE) %>% I()`-year-old should probably be removed.

```{r}
max_age <- 110
pool %>% filter(NUMERIC_AGE <= max_age) %>%
  scatterplotter(x = 'NUMERIC_AGE', y = 'LARGEST_GIFT_OR_PAYMENT'
                 , color = 'MG_PR_MODEL_DESC', ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(NUMERIC_AGE)), color = 'blue', linetype = 'dashed') +
  labs(title = bquote('Age versus largest gift' ~ (age <= .(max_age)) ))
```

For visits, based on the above exploration visit count needs a transformation.

```{r}
pool %>%
  scatterplotter(x = 'VISIT_COUNT', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(VISIT_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Log-sqrt visit count versus campaign giving') +
  scale_x_sqrt()
```

```{r}
pool %>%
  scatterplotter(x = 'VISIT_COUNT', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(VISIT_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Log-sqrt visit count versus largest gift') +
  scale_x_sqrt()
```

```{r}
pool %>%
  scatterplotter(x = 'VISIT_COUNT', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(VISIT_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Log-log visit count versus campaign giving') +
  scale_x_continuous(trans = 'log10plus1', breaks = c(0, 1, 10, 50, 100, 150, 200))
```

```{r}
pool %>%
  scatterplotter(x = 'VISIT_COUNT', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(VISIT_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Log-log visit count versus largest gift') +
  scale_x_continuous(trans = 'log10plus1', breaks = c(0, 1, 10, 50, 100, 150, 200))
```

The log-log plots look quite good. Visits $\geq$ 100 are outliers, but at first glance don't appear influential on the log-log scale (easy to test with e.g. Cook's D).

```{r}
pool %>%
  scatterplotter(x = 'YEARS_OF_GIVING', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(YEARS_OF_GIVING)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Years of giving versus campaign giving') +
  scale_x_sqrt()
```

```{r}
pool %>%
  scatterplotter(x = 'YEARS_OF_GIVING', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(YEARS_OF_GIVING)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Years of giving versus largest gift') +
  scale_x_sqrt()
```

The square root transformation does well for campaign giving, but not as well for largest gift or payment.

```{r}
# Box-Cox test for transformations
boxcox_lambdas <- seq(-1, 1, by = .01)
boxcox_lm <- lm(I(LARGEST_GIFT_OR_PAYMENT + 1) ~ YEARS_OF_GIVING, data = pool) %>%
  MASS::boxcox(lambda = boxcox_lambdas, plotit = FALSE)
maxLL <- boxcox_lm$x[which(boxcox_lm$y == max(boxcox_lm$y))]

# Plot results
boxcox_lm %>%
  unlist() %>%
  matrix(nrow = length(boxcox_lambdas)) %>%
  data.frame() %>%
  select(x = X1, y = X2) %>%
  ggplot(aes(x = x, y = y)) +
  geom_line() +
  geom_vline(aes(xintercept = x[which(y == max(y))]), color = 'blue', linetype = 'dashed') +
  labs(title = 'Box-Cox analysis', x = expression(lambda), y = 'log Likelihood')
```

$\lambda =$ `r maxLL %>% I()` is pretty close to a log transformation.

```{r}
# Best Box-Cox transformation, adding 1 so the response variable is strictly positive
boxcoxbest_trans <- function(x) {
  scales::trans_new(
    'boxcoxbest'
    , transform = function(x) {(x + 1)^maxLL}
    , inverse = function(x) {(x + 1)^(1/maxLL)}
  )
}
```

```{r}
grid.arrange(
# Plot Box-Cox results
  pool %>%
    scatterplotter(x = 'YEARS_OF_GIVING', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                   , ytrans = 'log10plus1', ylabels = scales::dollar) +
    geom_vline(aes(xintercept = mean(YEARS_OF_GIVING)), color = 'blue', linetype = 'dashed') +
    labs(title = 'Years of giving versus campaign giving, Box-Cox transformation', y = 'Largest gift') +
    scale_x_continuous(trans = 'boxcoxbest', breaks = c(0, 10))
# Plot log10 results
  , pool %>%
    scatterplotter(x = 'YEARS_OF_GIVING', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                   , ytrans = 'log10plus1', ylabels = scales::dollar) +
    geom_vline(aes(xintercept = mean(YEARS_OF_GIVING)), color = 'blue', linetype = 'dashed') +
    labs(title = 'Years of giving versus campaign giving, log transformation', y = 'Largest gift') +
    scale_x_continuous(trans = 'log10plus1')
)
```

After all that neither look linear, though they do look nearly linear around the mean.

```{r}
pool %>% mutate(last_3_yrs = factor(YEARS_OF_GIVING_LAST_3)) %>%
  ggplot(aes(x = last_3_yrs, y = CAMPAIGN_NEWGIFT_CMIT_CREDIT, color = MG_PR_MODEL_DESC)) +
  geom_boxplot() +
  facet_grid(. ~ MG_PR_MODEL_DESC) +
  scale_y_continuous(trans = 'log10plus1', labels = scales::dollar, breaks = 10^(0:20)) +
  labs(title = 'Years of giving of last 3 versus campaign giving')
```

```{r}
pool %>% mutate(last_3_yrs = factor(YEARS_OF_GIVING_LAST_3)) %>%
  ggplot(aes(x = last_3_yrs, y = LARGEST_GIFT_OR_PAYMENT, color = MG_PR_MODEL_DESC)) +
  geom_boxplot() +
  facet_grid(. ~ MG_PR_MODEL_DESC) +
  scale_y_continuous(trans = 'log10plus1', labels = scales::dollar, breaks = 10^(0:20)) +
  labs(title = 'Years of giving of last 3 versus largest gift')
```

I see a main effect for campaign giving.

```{r}
pool %>%
  scatterplotter(x = 'MG_250K_COUNT', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(MG_250K_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Count of $250K+ gifts versus campaign giving') +
  scale_x_sqrt()
```

```{r}
pool %>%
  scatterplotter(x = 'MG_250K_COUNT', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(MG_250K_COUNT)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Count of $250K+ gifts versus largest gift') +
  scale_x_sqrt()
```

As seen above, pretty much all of the observations are outliers. Might make sense to leave this one discretized.

```{r}
pool %>%
  scatterplotter(x = 'SEASON_TICKET_YEARS', y = 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(SEASON_TICKET_YEARS)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Season ticket years versus campaign giving') +
  scale_x_continuous(breaks = seq(0, 100, by = 2))
```

```{r}
pool %>%
  scatterplotter(x = 'SEASON_TICKET_YEARS', y = 'LARGEST_GIFT_OR_PAYMENT', color = 'MG_PR_MODEL_DESC'
                 , ytrans = 'log10plus1', ylabels = scales::dollar) +
  geom_vline(aes(xintercept = mean(SEASON_TICKET_YEARS)), color = 'blue', linetype = 'dashed') +
  labs(title = 'Season ticket years versus largest gift') +
  scale_x_continuous(breaks = seq(0, 100, by = 2))
```

This looks fine.

## Data exploration takeaways

  * Missing age values can be imputed as mean age
  * The 113-year-old is an outlier
  * Visit count could use log transform
  * Years of giving could use sqrt transform
  * \$250K+ gifts is a binary indicator

# Modeling ground rules

I'll outline some basic methodology in advance of modeling to try to account for [researcher degrees of freedom](http://journals.sagepub.com/doi/full/10.1177/0956797611417632).

As seen above, both campaign giving and largest lifetime gift have a strong linear relationship with the cultivation score, formulated as a 0 to `r max(pool$CULTIVATION_SCORE) %>% I()` point scale. Currently, all criteria count as a single point (equally weighted) but I suspect they impact actual giving behavior differently. I propose four models:

  1) Linear regression of campaign giving on cultivation score variables
  2) Linear regression of campaign on all available explanatory variables, essentially looking at the cultivation score variables controlling for the preexisting modeled scores
  3) Linear regression of largest cash transaction on cultivation score variables
  4) Linear regression of largest cash transaction on all available explanatory variables, again to control for the preexisting modeled scores

If a given explanatory variable has the same coefficient sign and similar nonzero magnitudes in all models, I'll take this as evidence of an association with giving behavior.

After withholding a random 20% of the data as a test set, I'll use ten-fold cross-validation to determine which variables should be included in each model. Model performance will be determined based on out-of-sample mean squared error, defined as usual:

$$ {\text{MSE}} = \frac{1}{n} \sum_{i=1}^{n} \left( Y_i - \hat{Y}_i \right)^2 $$

The 20% test set will be used to confirm the cross-validation results, and then final models will be constructed on the entire dataset for comparison.

## Data cleanup

First, perform some quick data clean-up.

```{r}
mdat <- pool %>% mutate(
  # Impute missing ages as mean age
  NUMERIC_AGE = case_when(
    !is.na(NUMERIC_AGE) ~ NUMERIC_AGE
    , TRUE ~ mean(NUMERIC_AGE, na.rm = TRUE)
  )
  # Impute missing affinity scores as mean affinity score
  , AFFINITY_SCORE = case_when(
    !is.na(AFFINITY_SCORE) ~ AFFINITY_SCORE
    , TRUE ~ mean(AFFINITY_SCORE, na.rm = TRUE)
  )
  # Create null factor levels for the MG_ID and MG_PR models
  , MG_ID_MODEL_DESC = fct_explicit_na(MG_ID_MODEL_DESC, 'Unscored') %>% fct_relevel('Unscored')
  , MG_PR_MODEL_DESC = fct_explicit_na(MG_PR_MODEL_DESC, 'Unscored') %>% fct_relevel('Unscored')
  # Create row numbers
  , rownum = 1:nrow(pool)
) %>% select(
  # Drop unhelpful fields
  -ID_NUMBER, -PROSPECT_ID, -PROSPECT_NAME, -NU_DEG, -NU_DEG_SPOUSE, -POTENTIAL_INTEREST_AREAS
  , -PREF_NAME_SORT, -MG_ID_MODEL_YEAR, -MG_ID_MODEL_SCORE, -MG_PR_MODEL_YEAR, -MG_PR_MODEL_SCORE
)
```

I'll use my `wranglR::KFoldXVal` function to create the cross-validation groups.

```{r}
k <- 11
# Create k groups: first is 20% of the data (prop = .2) and the others are equally sized
xval_inds <- pool %>% wranglR::KFoldXVal(k = k, prop = .2, seed = 12644)
# Remove outliers
outliers <- which(mdat$NUMERIC_AGE > max_age)
for (i in 1:k) {
  rm_idx <- which(xval_inds[[i]] %in% outliers)
  if (length(rm_idx) > 0) {xval_inds[[i]] <- xval_inds[[i]][-rm_idx]}
}
# Create groups
oos_inds <- xval_inds[[1]]
xval_inds <- xval_inds[2:k]
# Results
c(list(oos_inds), xval_inds) %>% summary
```

Group 1 is the out-of-sample validation set, and the other `r {k - 1} %>% I()` will be used for cross-validation.

Finally, I'll create a quick function to compute MSE.

```{r}
calc_mse <- function(y, yhat) {
  mean(
    (y - yhat)^2, na.rm = TRUE
  )
}
```

## Visualization functions

```{r}
# Create coefficients data frame
create_coefs <- function(model_list) {
  foreach(i = 1:length(model_list), .combine = 'rbind') %do% {
    tmp <- summary(model_list[[i]])$coefficients
    data.frame(tmp) %>%
    mutate(
      variable = rownames(tmp)
      , model = i
    ) %>% select(
      model
      , variable
      , beta.hat = Estimate
      , SE = Std..Error
      , t.val = t.value
      , Pr.t = Pr...t..
    ) %>% return()
  } %>% return()
}

# Plot R-squared
plot_r2 <- function(model_list, type = 'r.squared') {
  parser <- function(x) {
    tmpsum <- summary(x)
    paste0('tmpsum$', type) %>% parse(text = .) %>% eval() %>% return()
  }
  model_list %>%
  lapply(., function(x) parser(x)) %>% unlist() %>% data.frame(r.squared = .) %>%
    ggplot(aes(x = r.squared)) + 
    geom_density() +
    geom_vline(aes(xintercept = mean(r.squared)), color = 'blue', linetype = 'dashed', alpha = .5) +
    geom_rug(color = 'blue') +
    labs(title = bquote('Density plot of' ~ r^2 ~ 'results, mean' ~
        .(lapply(model_list, function(x) {summary(x)$r.squared}) %>% unlist() %>% mean() %>% round(3))
      )
      , x = bquote(r^2)
    )
}

# Plot cross-validated coefficients
plot_coefs <- function(model_list, conf_interval = 1 - p.sig) {
  crit_val <- qnorm({1 - conf_interval} / 2) %>% abs()
  coefs <- create_coefs(model_list)
  coefs %>% full_join(
    coefs %>% group_by(variable) %>%
      summarise(group.mean = mean(beta.hat), group.sd = sd(beta.hat))
    , by = c('variable', 'variable')
  ) %>%
  ggplot(aes(x = variable, y = beta.hat, color = factor(model))) +
  geom_segment(
    aes(
      xend = variable
      , y = group.mean - 2 * crit_val * group.sd
      , yend = group.mean + 2 * crit_val * group.sd
    ), color = 'gray', alpha = .25, size = 2) +
  geom_point() +
  geom_hline(yintercept = 0, alpha = .5) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(
    title = 'Coefficient estimates per cross-validation model'
    , y = bquote(hat(beta))
    , color = 'cross-validation sample'
  )
}

# Table of coefficient +/- counts
coef_pm_table <- function(model_list, pval) {
  create_coefs(model_list) %>%
  group_by(variable) %>%
  summarise(
    `+` = sum(sign(beta.hat) == 1 & Pr.t < pval)
    , `0` = sum(Pr.t >= pval)
    , `-` = sum(sign(beta.hat) < 0 & Pr.t < pval)
  )
}

# Compute predictions
calc_preds <- function(model_list, xval, yname) {
  yhats <- list()
  for (i in 1:length(model_list)) {
    yhats[[i]] <- data.frame(
      model = i
      , row = xval[[i]]
      , preds = model_list[[i]] %>% predict(newdata = mdat[xval[[i]], ])
      , truth = mdat[xval[[i]], yname] %>% unlist() %>% log10plus1()
    )
  }
  return(yhats)
}
calc_outsample_mse <- function(model_list, xval, yname) {
  calc_preds(model_list, xval, yname) %>%
    lapply(function(x) calc_mse(y = x$truth, yhat = x$preds)) %>%
    unlist()
}

# Plot MSEs by insample/outsample
plot_mses <- function(model_list, xval, truth) {
  mses <- data.frame(
    insample = model_list %>%
      lapply(function(x) calc_mse(y = model.frame(x)[, 1], yhat = predict(x))) %>%
      unlist()
    , outsample = calc_outsample_mse(model_list, xval, truth)
  ) %>% gather('type', 'MSE', 1:2)
  mses %>%
    ggplot(aes(x = MSE, color = type)) +
    geom_density() +
    geom_vline(
      xintercept = mses %>% filter(type == 'insample') %>% select(MSE) %>% unlist %>% mean()
      , color = 'red', linetype = 'dashed', alpha = .5
    ) +
    geom_vline(
      xintercept = mses %>% filter(type == 'outsample') %>% select(MSE) %>% unlist %>% mean()
      , color = 'darkcyan', linetype = 'dashed', alpha = .5
    ) +
    geom_rug() +
    labs(
      title = bquote('MSE across samples, means =' ~
          .(mses %>% group_by(type) %>% summarise(mean = mean(MSE)) %>%
              select(mean) %>% unlist() %>% round(3) %>% paste(collapse = ', ')
          )
        )
    )
}

# Merges predicted results into one large data frame each for insample and outsample
calc_resids <- function(model_list, xval, yname) {
  insample <- foreach(i = 1:length(model_list), .combine = 'rbind') %do% {
    data.frame(
      model = i
      , preds = model_list[[i]] %>% predict()
      , truth = model.frame(model_list[[i]])[, 1]
    ) %>% mutate(
      residuals = truth - preds
    )
  }
  preds <- calc_preds(model_list, xval, yname)
  outsample <- foreach(i = 1:length(model_list), .combine = 'rbind') %do% {
    preds[[i]]
  } %>% mutate(
    residuals = truth - preds
  )
  return(list(insample = insample, outsample = outsample))
}

# Plot standardized residuals; returns a list of ggplot objects $insample and $outsample
plot_resids <- function(model_list, xval, yname, filter = 'TRUE') {
  resids <- calc_resids(model_list, xval, yname)
  # Plot residuals vs fitted for in-sample data
  insample <- resids$insample %>% filter_(filter) %>%
    ggplot(aes(x = preds, y = residuals, color = factor(model))) +
    geom_point(alpha = .01) +
    geom_smooth(se = FALSE) +
    labs(title = 'In-sample residuals versus fitted', color = 'cross-validation sample')
  # Plot residuals vs fitted for out-of-sample data
  outsample <- resids$outsample %>% filter_(filter) %>%
    ggplot(aes(x = preds, y = residuals, color = factor(model))) +
    geom_point(alpha = .1) +
    geom_smooth(se = FALSE) +
    labs(title = 'Out-of-sample residuals versus fitted', color = 'cross-validation sample')
  return(list(insample = insample, outsample = outsample))
}

# Plot normal Q-Q visualization for residuals
plot_qq <- function(model_list, xval, yname, filter = 'TRUE') {
  resids <- calc_resids(model_list, xval, yname)
  # In-sample Q-Q plot with standardized residuals
  insample <- resids$insample %>% mutate(st.resid = residuals/sd(residuals)) %>% filter_(filter) %>%
    ggplot(aes(sample = st.resid, color = factor(model))) +
    geom_qq(alpha = .05) +
    geom_qq_line() +
    labs(title = 'In-sample Q-Q plot with standardized residuals'
         , color = 'cross-validation sample')
  # Out-of-sample Q-Q plot
  outsample <- resids$outsample %>% mutate(st.resid = residuals/sd(residuals)) %>% filter_(filter) %>%
    ggplot(aes(sample = st.resid, color = factor(model))) +
    geom_qq(alpha = .05) +
    geom_qq_line() +
    labs(title = 'Out-of-sample Q-Q plot with standardized residuals'
         , color = 'cross-validation sample')
  return(list(insample = insample, outsample = outsample))
}

# Compute partial residuals
calc_partial_resids <- function(model, inds, xname, df = NULL) {
  # Quick hack to get around update() environment, not really recommended
  scoped_xname <<- xname
  scoped_spline_df <<- df
  # Partial residuals data frame
  data.frame(
    x = mdat %>% filter(rownum %in% unlist(inds)) %>% select_(xname) %>% unlist()
    , y = model.frame(model)[, 1]
    , y.hat = if (!is.null(df)) {
      update(model, formula = . ~ ns(parse(text = scoped_xname) %>% eval(), df = scoped_spline_df)) %>% fitted()
    } else {
      update(model, formula = . ~ parse(text = scoped_xname) %>% eval()) %>% fitted()
    }
  ) %>% mutate(
    # Compute partial residuals
    y.partial.resid = if (!is.null(df)) {
      {model %>% residuals(type = 'partial')}[, paste0('ns(', parse(text = xname), ', df = ', df, ')')]
    } else {
      {model %>% residuals(type = 'partial')}[, xname]
    }
    # Regression on partial residuals
    , y.hat.partial = if (!is.null(df)) {
      lm(y.partial.resid ~ ns(x, df)) %>% fitted()
    } else {
      lm(y.partial.resid ~ x) %>% fitted()
    }
  ) %>% return()
}
```

# Campaign linear models

## Cultivation score predictors {#campaign-cultivation-score-predictors}

Regress campaign giving against each of the cultivation score predictors.

```{r}
# List to store campaign linear models
clms <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  clms[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
        ACTIVE_PROPOSALS +
        AGE +
        PM_VISIT_LAST_2_YRS +
        VISITS_5PLUS +
        AF_25K_GIFT +
        GAVE_IN_LAST_3_YRS +
        MG_250K_PLUS +
        PRESIDENT_VISIT +
        TRUSTEE_OR_ADVISORY_BOARD +
        Alumnus +
        DEEP_ENGAGEMENT +
        CHICAGO_HOME
      , data = .
    )
}
```

The full (and hard to read) results for each model are in the [appendix](#appendix-campaign-cultivation).

We can extract a few parameters of interest.

```{r}
plot_r2(clms) +
  geom_text(y = seq(10, 100, length.out = k - 1), label = 1:(k - 1), color = 'blue') +
  xlim(c(.45, .48))
```

The average $r^2 =$ `r lapply(clms, function(x) {summary(x)$r.squared}) %>% unlist() %>% mean() %>% round(3) %>% I()` is quite a good result.

```{r, fig.height = 6}
p.sig <- 1E-2
plot_coefs(clms)
```

```{r, rows.print = 100}
coef_pm_table(clms, p.sig)
```

The coefficients are extremely tightly clustered within each cross-validation set. Interestingly, age and alumni status both have negative coefficients. All are significant at $p =$ `r p.sig %>% I()`.

Here are the actual prediction MSEs:

```{r}
plot_mses(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')
```

As usual, in-sample performance is moderately optimistic.

```{r}
plot_resids(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$insample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

```{r}
plot_resids(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

At first glance, the linear trend evident in these results is concerning. Upon second glance, while still sobering it really just reinforces what we'd observed in the first of the two plots in the [Background](#background) section: campaign giving is decidedly nonlinear as cultivation scores approach their high/low limits. As seen above, the underlying factors should be transformed, and possibly modeled nonlinearly (splines).


```{r}
plot_qq(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$insample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

```{r}
plot_qq(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

The quantile-quantile plot further illustrates the issue. The drift above the reference line to the left and below it to the right suggests less density than expected in the tails, which follows given that the range is bound -- it's not possible to give less than a nondonor or (for practical purposes) more than a 9-figure donor.

The story is completely different when looking only at those who actually gave:


```{r}
plot_qq(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', filter = 'truth > 0')$insample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```


```{r}
plot_qq(clms, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT', filter = 'truth > 0')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

This is quite a bit nicer, suggesting only slight skewness in the tails.

**Takeaway.** When modeling giving amounts in the future, consider fitting a conditional model. Something along the lines of:

$$ E(\text{giving amount} ~ | ~ \text{donor status} = 1) $$

## All predictors 1 {#campaign-all-predictors-1}

The second model compares the variables underlying each PG score indicator, plus supplemental predictors. The first model incudes everything and I'll prune from there using MSE. Initial spline df are fairly arbitrary.

```{r}
# List to store campaign linear models
clmaps <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  clmaps[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
        ACTIVE_PROPOSALS +
        ns(NUMERIC_AGE, df = 5) + # Underlying variable to AGE indicator
        PM_VISIT_LAST_2_YRS +
        log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
        AF_25K_GIFT +
        sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
        ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
        MG_250K_PLUS + # Decided to leave as factor
        PRESIDENT_VISIT +
        TRUSTEE_OR_ADVISORY_BOARD +
        Alumnus +
        DOUBLE_ALUM + # Deep Engagement component
        EVER_PARENT + # Deep Engagement component
        SEASON_TICKET_YEARS + # Deep Engagement component
        CHICAGO_HOME +
        QUAL_LEVEL +
        AFFINITY_SCORE +
        MG_PR_MODEL_DESC
      , data = .
    )
}
```

Full results are in the [appendix](#appendix-campaign-all-predictors-1).

```{r}
plot_r2(clmaps) +
  geom_text(y = seq(10, 150, length.out = k - 1), label = 1:(k-1), color = 'blue') +
  xlim(c(.72, .735))
```

This is a much higher $r^2$ than seen above, but the model also includes many more predictors. Consider the MSE.

```{r}
plot_mses(clmaps, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')
```

Well, that's pretty conclusive -- this is also much lower than that seen previously. This implies that on average the predicted giving amount is less than a factor of 10 off. Which predictors contribute to the performance?

```{r, fig.height = 6}
plot_coefs(clmaps)
```

```{r, rows.print = 100}
coef_pm_table(clmaps, p.sig)
```

After accounting for the other variables, things that don't seem to matter include active proposals, Chicago home address, two of the deep engagement indicators, president visits, qualification level (inconsistent), and committee participation. Note that the "Future Prospect" factor level only appears 9 times -- one of the cross-validation samples must not have had anyone rated at that level. Additionally, some of these are likely already included in the affinity score.

```{r}
plot_resids(clmaps, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2))
```

```{r}
plot_qq(clmaps, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample 
```

This already looks a lot nicer than the [Cultivation score predictors](#campaign-cultivation-score-predictors) model! Now let's try again, dropping the less-interesting predictors.

## All predictors 2 {#campaign-all-predictors-2}

Full results are in the [appendix](#appendix-campaign-all-predictors-2).

```{r}
# List to store campaign linear models
clmaps2 <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  clmaps2[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
        ns(NUMERIC_AGE, df = 5) + # Underlying variable to AGE indicator
        PM_VISIT_LAST_2_YRS +
        log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
        AF_25K_GIFT +
        sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
        ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
        MG_250K_PLUS + # Decided to leave as factor
        Alumnus +
        SEASON_TICKET_YEARS + # Deep Engagement component
        AFFINITY_SCORE +
        MG_PR_MODEL_DESC
      , data = .
    )
}
```

```{r}
plot_mses(clmaps2, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')
```

That's a marginal increase in outsample MSE (`r calc_outsample_mse(clmaps, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT') %>% mean() %>% round(4)` to `r calc_outsample_mse(clmaps2, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT') %>% mean() %>% round(4)`, difference of `r scales::percent({calc_outsample_mse(clmaps2, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT') %>% mean()}/{calc_outsample_mse(clmaps, xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT') %>% mean()} - 1)`) with many fewer predictors.

```{r, fig.height = 6}
plot_coefs(clmaps2)
```

```{r, rows.print = 100}
coef_pm_table(clmaps2, p.sig)
```

The spline degrees of freedom for numeric age could be tweaked.

## All predictors splines test {#campaign-all-predictors-splines}

```{r}
splines_max <- 10
clmaps3 <- list()
for (s in 1:splines_max) {
  clmaps3[[s]] <- list()
  for(i in 1:length(xval_inds)) {
    # Create linear model excluding the holdout and out-of-sample indices
    clmaps3[[s]][[i]] <- mdat %>%
      filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
      lm(
        log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
          ns(NUMERIC_AGE, df = s) + # Underlying variable to AGE indicator
          PM_VISIT_LAST_2_YRS +
          log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
          AF_25K_GIFT +
          sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
          ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
          MG_250K_PLUS + # Decided to leave as factor
          Alumnus +
          SEASON_TICKET_YEARS + # Deep Engagement component
          AFFINITY_SCORE +
          MG_PR_MODEL_DESC
        , data = .
      )
  }
}
```

Try different spline degrees of freedom for NUMERIC_AGE. Full results are in the [appendix](#appendix-campaign-all-predictors-splines).

Consider the distribution of MSEs for each model.

```{r}
# Calculate MSEs for each model
spline_mse <- foreach(s = 1:splines_max, .combine = rbind) %do% {
  mses <- calc_outsample_mse(clmaps3[[s]], xval_inds, 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')
  data.frame(spline.df = s, xv_group = factor(1:(k-1)), mses)
}
```
```{r}
# Plot results
spline_mse %>%
  ggplot(aes(x = spline.df, y = mses)) +
  geom_point(aes(color = xv_group)) +
  geom_smooth() +
  scale_x_continuous(breaks = 1:splines_max, minor_breaks = NULL) +
  labs(x = 'spline df', y = 'MSE', color = 'cross-validation sample')
```

For practical purposes there's not much difference between the different choices. It looks like 4 or 5 is where the mean MSE levels out, so I'll stick with my initial choice.

## Comparison {#campaign-comparison}

Create the two final models and check them on out-of-sample data.

```{r}
# Predict campaign giving, PG score indicators
clm_final <- clms[[1]] %>% update(data = mdat %>% filter(rownum %in% unlist(xval_inds)))

# Predict campaign giving, underlying factors
clmap_final <- clmaps2[[1]] %>% update(data = mdat %>% filter(rownum %in% unlist(xval_inds)))
```

Full results in the [appendix](#appendix-campaign-final-models)

```{r}
plot_resids(list(clmap_final, clm_final), list(oos_inds, oos_inds), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  scale_color_discrete(labels = c('All predictors', 'PG indicators')) +
  labs(color = 'Model')
```

The "All predictors" residuals look better than the "PG indicators"" residuals.

```{r}
plot_qq(list(clmap_final, clm_final), list(oos_inds, oos_inds), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  scale_color_discrete(labels = c('All predictors', 'PG indicators')) +
  labs(color = 'Model')
```

Note the reference lines -- "All predictors" is much closer to a normal distribution, albeit with some positive skewness.

```{r}
c_mses_final <- rbind(
  calc_preds(list(clm_final), list(oos_inds), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')[[1]] %>%
    mutate(model = 'PG indicators')
  , calc_preds(list(clmap_final), list(oos_inds), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')[[1]] %>%
    mutate(model = 'All predictors')
) %>% mutate(
  model = factor(model)
  , sq.error = (truth - preds)^2
)
c_mses_final <- c_mses_final %>% left_join(
  c_mses_final %>% group_by(model) %>% summarise(mse = mean(sq.error))
  , by = c('model', 'model')
)
c_mses_final %>%
  ggplot(aes(x = sq.error, color = model)) +
  geom_density() +
  geom_vline(aes(xintercept = mean(sq.error)), linetype = 'dotted', alpha = .5) +
  geom_vline(aes(xintercept = mse, color = model), linetype = 'dashed') +
  xlim(c(0, 5)) +
  labs(
    x = 'squared error'
    , title = bquote('Squared error across models, means = ' ~
        .(c_mses_final %>% group_by(model) %>% summarise(mse = mean(mse)) %>% select(mse) %>%
            unlist() %>% round(3) %>% paste(collapse = ', ')))
  )
```

We never want to see bimodal squared errors -- it looks like "All predictors" is a much better regression model. These are all very close to the results seen in the corresponding sections above, which is reassuring.

Finally, let's look at the impact of age in the splines model. I'll try plotting against both the raw $\textbf{y}$ and the partial residuals, which are obtained by removing the effects of all the predictors besides the $\textbf{x}_j$ we're interested in, e.g.

$$ \boldsymbol{\hat{\epsilon}} = \textbf{y} - \hat{\textbf{y}} = \textbf{y} - \sum_{i} {\textbf{x}_i \hat{\boldsymbol{\beta}}_i} $$
$$ \boldsymbol{\hat{\epsilon}}_\text{partial} = \textbf{y} - \sum_{i \neq j} {\textbf{x}_i \hat{\boldsymbol{\beta}}_i} $$

```{r}
calc_partial_resids(clmap_final, xval_inds, 'NUMERIC_AGE', 5) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 2, color = 'darkgray') +
  geom_line(aes(y = y.hat), color = 'red') +
  labs(title = 'Numeric age spline versus campaign giving', x = 'age', y = bquote(log[10] ~ 'campaign giving'))
```

```{r}
calc_partial_resids(clmap_final, xval_inds, 'NUMERIC_AGE', 5) %>%
  ggplot(aes(x = x, y = y.partial.resid)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  geom_line(aes(y = y.hat.partial), color = 'red') +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  labs(title = 'Numeric age spline versus partial residuals for campaign giving'
       , x = 'age'
       , y = 'partial residuals')
```

Recall that in the PG indicators model, age had a negative coefficient. This paints a more nuanced picture. The first plot regresses campaign giving on the natural spline of age and shows a dip in expected giving for 60-year-olds, a gradual increase to age 80 or so, and then a slow decline. However, we know from previous experience that age is closely correlated with many other predictors of giving (years of giving, total giving, capacity evaluation, and so on). The partial residual plot corrects for this by removing the effect of all the other predictors and regressing the resulting residuals on age. Now, controlling for the other variables, expected campaign giving appears to reach its maximum in the late 40s or early 50s, and decreases thereafter.

## Thoughts

Compare the coefficients for both models.

```{r, rows.print = 100, warning = FALSE, message = FALSE}
full_join(
  data.frame(var = coef(clm_final) %>% names, pg.inds.model = coef(clm_final))
  , data.frame(var = coef(clmap_final) %>% names, all.predictors.model = coef(clmap_final))
) %>% mutate(
  # Factor levels in order we want them to appear in below table
  var = factor(var, levels = c('(Intercept)', 'ACTIVE_PROPOSALS', 'AGE', 'ns(NUMERIC_AGE, df = 5)1'
    , 'ns(NUMERIC_AGE, df = 5)2', 'ns(NUMERIC_AGE, df = 5)3', 'ns(NUMERIC_AGE, df = 5)4'
    , 'ns(NUMERIC_AGE, df = 5)5', 'PRESIDENT_VISIT', 'PM_VISIT_LAST_2_YRS', 'VISITS_5PLUS'
    , 'log10plus1(VISIT_COUNT)', 'AF_25K_GIFT', 'GAVE_IN_LAST_3_YRS', 'sqrt(YEARS_OF_GIVING)'
    , 'ns(YEARS_OF_GIVING_LAST_3, df = 2)1', 'ns(YEARS_OF_GIVING_LAST_3, df = 2)2', 'MG_250K_PLUS'
    , 'TRUSTEE_OR_ADVISORY_BOARD', 'Alumnus', 'DEEP_ENGAGEMENT', 'CHICAGO_HOME', 'SEASON_TICKET_YEARS'
    , 'AFFINITY_SCORE', 'MG_PR_MODEL_DESCBottom Tier', 'MG_PR_MODEL_DESCMiddle Tier', 'MG_PR_MODEL_DESCTop Tier'))
) %>% arrange(var)
```

  * The intercept for the all predictors model is much closer to 0. The relatively large intercept for the indicators model likely contributes to its odd residuals behavior (recall the bimodal squared errors).
  * Age has a nonlinear relationship with giving. The oldest donors are expected to give slightly less.
  * The impact of visits is much smaller for the all predictors model, though more visit activity is always associated with greater giving.
  * Giving appears to have a comparable impact between the two models, though interpretation is tricky given that the all predictors model splits its impact over more predictors.
  * Alumni status leads to *less* predicted giving in both models! While surprising on the face of it, this actually fits with what I've observed in other datasets. Consider that nonalumni are often added to the database only once they've already made a gift, demonstrating both affinity and philanthropic interest.
  * The affinity score and MG prioritization models work!
  
Overall, with Campaign giving as $y$, several of the PG model variables do seem to offer predictive power above and beyond whatever factors are rolled into the affinity score and MG prioritization models. In particular, age, PM visits, total visits, $25K+ AF gifts, total years of giving, MG gifts, alumni status, and season tickets were still statistically significant when including the other modeled scores. It may be worthwhile paying special attention to these characteristics.

# Transaction linear models

Repeat the above, but let $y_i$ be transformed largest gift or payment.

## Cultivation score predictors {#trans-cultivation-score-predictors}

Full summary in the [appendix](#appendix-trans-cultivation).

```{r}
# List to store transaction linear models
tlms <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  tlms[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(LARGEST_GIFT_OR_PAYMENT) ~
        ACTIVE_PROPOSALS +
        AGE +
        PM_VISIT_LAST_2_YRS +
        VISITS_5PLUS +
        AF_25K_GIFT +
        GAVE_IN_LAST_3_YRS +
        MG_250K_PLUS +
        PRESIDENT_VISIT +
        TRUSTEE_OR_ADVISORY_BOARD +
        Alumnus +
        DEEP_ENGAGEMENT +
        CHICAGO_HOME
      , data = .
    )
}
```

```{r}
plot_r2(tlms) +
  geom_text(y = seq(10, 200, length.out = k - 1), label = 1:(k - 1), color = 'blue') +
  xlim(c(.325, .335))
```

Here, the mean $r^2 =$ `r lapply(tlms, function(x) {summary(x)$r.squared}) %>% unlist() %>% mean() %>% round(3) %>% I()` is a bit lower than that seen in the campaign models.

```{r, fig.height = 6}
plot_coefs(tlms)
```

```{r, rows.print = 100}
coef_pm_table(tlms, p.sig)
```

Interestingly, all the coefficients are positive besides alumni status, which is not predictive! President visits are hit or miss, which follows given that the quality of visit data is time dependent while gift data is not. The rest all have positive coefficients.

```{r}
plot_mses(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')
```

This is actually much lower than what was seen in the campaign cultivation score model.

```{r}
plot_resids(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$insample
```

```{r}
plot_resids(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

This exhibits a similar downward trend as scene in the campaign cultivation score model, so I don't expect the Q-Q plots to be acceptable.

```{r}
plot_qq(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$insample
```

```{r}
plot_qq(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

That looks almost like a discontinuity at -1.

```{r}
mdat %>% mutate(x.trans = log10plus1(LARGEST_GIFT_OR_PAYMENT)) %>%
  ggplot(aes(x = x.trans)) +
  geom_histogram(alpha = .5, binwidth = .1) +
  geom_density(aes(y = ..count..)) +
  geom_vline(aes(xintercept = mean(x.trans)), color = 'blue', linetype = 'dashed') +
  geom_vline(aes(xintercept = mean(x.trans) - sd(x.trans)), color = 'blue', linetype = 'dotted')
```

As before, the combination of binary indicators and all those 0 observations seem to do a number on the model.

```{r}
plot_qq(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT', filter = 'truth > 0')$insample
```

```{r}
plot_qq(tlms, xval_inds, 'LARGEST_GIFT_OR_PAYMENT', filter = 'truth > 0')$outsample
```

But once again, the expected value conditioned on being a donor is nearly normal. This approach, $E(\text{giving} ~ | ~ \text{donor status} = 1)$, is definitely worth considering in the future.

## All predictors 1 {#trans-all-predictors-1}

Repeat the above with the underlying variables. Again, the first model will include all variables and I'll prune it back according to MSE and variable significance. Spline df are set as in the corresponding campaign model.

Full summary in the [appendix](#appendix-trans-all-1).

```{r}
# List to store transaction linear models
tlmaps <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  tlmaps[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(LARGEST_GIFT_OR_PAYMENT) ~
        ACTIVE_PROPOSALS +
        ns(NUMERIC_AGE, df = 5) + # Underlying variable to AGE indicator
        PM_VISIT_LAST_2_YRS +
        log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
        AF_25K_GIFT +
        sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
        ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
        MG_250K_PLUS + # Decided to leave as factor
        PRESIDENT_VISIT +
        TRUSTEE_OR_ADVISORY_BOARD +
        Alumnus +
        DOUBLE_ALUM + # Deep Engagement component
        EVER_PARENT + # Deep Engagement component
        SEASON_TICKET_YEARS + # Deep Engagement component
        CHICAGO_HOME +
        QUAL_LEVEL +
        AFFINITY_SCORE +
        MG_PR_MODEL_DESC
      , data = .
    )
}
```

```{r}
plot_r2(tlmaps) +
  geom_text(y = seq(10, 250, length.out = k - 1), label = 1:(k-1), color = 'blue') +
  xlim(c(.593, .605))
```

```{r}
plot_mses(tlmaps, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')
```

While $r^2$ is not as high as that achieved by the campaign all predictors model, this is the lowest MSE seen yet.

```{r, fig.height = 6}
plot_coefs(tlmaps)
```

```{r, rows.print = 100}
coef_pm_table(tlmaps, p.sig)
```

Intersetingly, there are some notable differences compared to campaign giving. Here, having a Chicago home address does matter, as do parent affinity, and committee participation. Qualification level is still inconsistent, and age appears to matter much less here compared to the campaign model. Interestingly, compared to the previous model, alumni status is significant again once these underlying factors are included.

```{r}
plot_resids(tlmaps, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

```{r}
plot_qq(tlmaps, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

That residuals plot is still quite poor, but the Q-Q plot is respectable.

## All predictors 2 {#trans-all-predictors-2}

Re-fit the above, dropping qualification level, active proposals, and double alum. It also appears that fewer splie df for age are called for.

Full results in the [appendix](#appendix-trans-all-2).

```{r}
# List to store transaction linear models
tlmaps2 <- list()
for(i in 1:length(xval_inds)) {
  # Create linear model excluding the holdout and out-of-sample indices
  tlmaps2[[i]] <- mdat %>%
    filter(rownum %in% unlist(xval_inds)[unlist(xval_inds) %nin% xval_inds[[i]]]) %>%
    lm(
      log10plus1(LARGEST_GIFT_OR_PAYMENT) ~
        ns(NUMERIC_AGE, df = 3) + # Underlying variable to AGE indicator
        PM_VISIT_LAST_2_YRS +
        log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
        AF_25K_GIFT +
        sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
        ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
        MG_250K_PLUS + # Decided to leave as factor
        PRESIDENT_VISIT +
        TRUSTEE_OR_ADVISORY_BOARD +
        Alumnus +
        EVER_PARENT + # Deep Engagement component
        SEASON_TICKET_YEARS + # Deep Engagement component
        CHICAGO_HOME +
        AFFINITY_SCORE +
        MG_PR_MODEL_DESC
      , data = .
    )
}
```

```{r}
plot_r2(tlmaps2) +
  geom_text(y = seq(10, 250, length.out = k - 1), label = 1:(k-1), color = 'blue') +
  xlim(c(.59, .60))
```

```{r}
plot_mses(tlmaps2, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')
```

Similar $r^2$ and error, so the smaller model is preferred.

```{r, fig.height = 6}
plot_coefs(tlmaps2)
```


```{r, rows.print = 100}
coef_pm_table(tlmaps2, p.sig)
```

Alumni and parent affiliation *reduce* the expected largest gift amount, on average.

```{r}
plot_resids(tlmaps2, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

```{r}
plot_qq(tlmaps2, xval_inds, 'LARGEST_GIFT_OR_PAYMENT')$outsample
```

Still can't seem to fix those residuals.

## Comparison {#trans-comparison}

Check the models on out-of-sample data. Full results in the [appendix](#appendix-trans-final-models).

```{r}
# Predict largest transaction, PG score indicators
tlm_final <- tlms[[1]] %>% update(data = mdat %>% filter(rownum %in% unlist(xval_inds)))
# Predict largest transaction, underlying factors
tlmap_final <- tlmaps2[[1]] %>% update(data = mdat %>% filter(rownum %in% unlist(xval_inds)))
```

```{r}
plot_resids(list(tlmap_final, tlm_final), list(oos_inds, oos_inds), 'LARGEST_GIFT_OR_PAYMENT')$outsample +
  scale_x_continuous(breaks = seq(0, 10, by = 2)) +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  scale_color_discrete(labels = c('All predictors', 'PG indicators')) +
  labs(color = 'Model')
```

Honestly, neither model looks great.

```{r}
plot_qq(list(tlmap_final, tlm_final), list(oos_inds, oos_inds), 'LARGEST_GIFT_OR_PAYMENT')$outsample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  scale_color_discrete(labels = c('All predictors', 'PG indicators')) +
  labs(color = 'Model')
```

At least the the all predictors residuals lie closer to a normal distribution.

```{r}
t_mses_final <- rbind(
  calc_preds(list(tlm_final), list(oos_inds), 'LARGEST_GIFT_OR_PAYMENT')[[1]] %>%
    mutate(model = 'PG indicators')
  , calc_preds(list(tlmap_final), list(oos_inds), 'LARGEST_GIFT_OR_PAYMENT')[[1]] %>%
    mutate(model = 'All predictors')
) %>% mutate(
  model = factor(model)
  , sq.error = (truth - preds)^2
)
t_mses_final <- t_mses_final %>% left_join(
  t_mses_final %>% group_by(model) %>% summarise(mse = mean(sq.error))
  , by = c('model', 'model')
)
t_mses_final %>%
  ggplot(aes(x = sq.error, color = model)) +
  geom_density() +
  geom_vline(aes(xintercept = mean(sq.error)), linetype = 'dotted', alpha = .5) +
  geom_vline(aes(xintercept = mse, color = model), linetype = 'dashed') +
  xlim(c(0, 5)) +
  labs(
    x = 'squared error'
    , title = bquote('Squared error across models, means = ' ~
        .(t_mses_final %>% group_by(model) %>% summarise(mse = mean(mse)) %>% select(mse) %>%
            unlist() %>% round(3) %>% paste(collapse = ', ')))
  )
```

Also very close to the previous results. Again, all predictors is a better regression model.

Finally, consider the effect of age on largest gift.

```{r}
calc_partial_resids(tlmap_final, xval_inds, 'NUMERIC_AGE', 3) %>%
  ggplot(aes(x = x, y = y)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 2, color = 'darkgray') +
  geom_line(aes(y = y.hat), color = 'red') +
  labs(title = 'Numeric age spline versus largest transaction', x = 'age', y = bquote(log[10] ~ 'largest transaction'))
```

```{r}
calc_partial_resids(tlmap_final, xval_inds, 'NUMERIC_AGE', 3) %>%
  ggplot(aes(x = x, y = y.partial.resid)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  geom_line(aes(y = y.hat.partial), color = 'red') +
  labs(title = 'Numeric age spline versus partial residuals for largest transaction'
       , x = 'age'
       , y = 'partial residuals')
```

I find this surprising. Unlike in the [campaign model comparison](#campaign-comparison), largest gift size appears to increase monotonically with age. However, controlling for the other variables, we again see peak giving occurs in the late 40s or early 50s, though the fitted line is nearly flat (confirming the lesser impact of age in this model compared to the previous one). Apparently there are very few large estate gifts compared to the number of older donors.

**Takeaway.** Today's age is not particularly predictive of largest gift; in future models consider age at time of gift instead. This could also be used to predict next gift size.

## Thoughts

Here are the coefficients for each model.

```{r, rows.print = 100, warning = FALSE, message = FALSE}
full_join(
  data.frame(var = coef(tlm_final) %>% names, pg.inds.model = coef(tlm_final))
  , data.frame(var = coef(tlmap_final) %>% names, all.predictors.model = coef(tlmap_final))
) %>% mutate(
  # Factor levels in order we want them to appear in below table
  var = factor(var, levels = c('(Intercept)', 'ACTIVE_PROPOSALS'
    , 'AGE', 'ns(NUMERIC_AGE, df = 3)1', 'ns(NUMERIC_AGE, df = 3)2', 'ns(NUMERIC_AGE, df = 3)3'
    , 'PRESIDENT_VISIT', 'PM_VISIT_LAST_2_YRS', 'VISITS_5PLUS', 'log10plus1(VISIT_COUNT)'
    , 'AF_25K_GIFT', 'GAVE_IN_LAST_3_YRS', 'sqrt(YEARS_OF_GIVING)'
    , 'ns(YEARS_OF_GIVING_LAST_3, df = 2)1', 'ns(YEARS_OF_GIVING_LAST_3, df = 2)2', 'MG_250K_PLUS'
    , 'TRUSTEE_OR_ADVISORY_BOARD'
    , 'Alumnus', 'EVER_PARENT', 'DEEP_ENGAGEMENT', 'CHICAGO_HOME', 'SEASON_TICKET_YEARS'
    , 'AFFINITY_SCORE', 'MG_PR_MODEL_DESCBottom Tier', 'MG_PR_MODEL_DESCMiddle Tier', 'MG_PR_MODEL_DESCTop Tier'))
) %>% arrange(var)
```

  * The intercept is smaller in the all predictors model.
  * Controlling for the other predictors, the effect of age is fairly minor.
  * Unsurprisingly, past major giving is the strongest predictor of...major gift behavior.
  * Again, the affinity score and prioritization models work -- though here, bottom and middle tier don't appear to have given at different levels in the past (which is likely why they weren't top tier).

All of the PG indicators apparently help estimate largest gift amount, above and beyond the information that other variables and modeled scores provide. Importantly, with the exception of alumni status, all the signs point in the same direction.

# Final Comparison {#final-comparison}

Rebuild all models using the entire dataset. Full results are in the [appendix](#appendix-final-models).

```{r, warning = FALSE, message = FALSE}
# Final models list
final_models <- list(
  campaign.pg.score = clm_final %>% update(data = mdat %>% filter(rownum %nin% outliers))
  , largest.trans.pg.score = tlm_final %>% update(data = mdat %>% filter(rownum %nin% outliers))
  , campaign.all.preds = clmap_final %>% update(data = mdat %>% filter(rownum %nin% outliers))
  , largest.trans.all.preds = tlmap_final %>% update(data = mdat %>% filter(rownum %nin% outliers))
)

# Coefficients data frame
final_coefs <- full_join(
  data.frame(var = coef(final_models[[1]]) %>% names()
             , campaign.pg = coef(final_models[[1]]))
  , data.frame(var = coef(final_models[[2]]) %>% names()
               , largest.pg = coef(final_models[[2]]))
) %>% full_join(
  data.frame(var = coef(final_models[[3]]) %>% names() %>% str_replace(', df = [0-9]*', '')
             , campaign.all = coef(final_models[[3]]))
) %>% full_join(
  data.frame(var = coef(final_models[[4]]) %>% names() %>% str_replace(', df = [0-9]*', '')
             , largest.all = coef(final_models[[4]]))
) %>% mutate(
  # Factor levels in order we want them to appear in below table
  var = factor(var, levels = c('(Intercept)', 'ACTIVE_PROPOSALS'
    , 'AGE', 'ns(NUMERIC_AGE)1', 'ns(NUMERIC_AGE)2', 'ns(NUMERIC_AGE)3', 'ns(NUMERIC_AGE)4', 'ns(NUMERIC_AGE)5'
    , 'PRESIDENT_VISIT', 'PM_VISIT_LAST_2_YRS', 'VISITS_5PLUS', 'log10plus1(VISIT_COUNT)'
    , 'AF_25K_GIFT', 'GAVE_IN_LAST_3_YRS', 'sqrt(YEARS_OF_GIVING)'
    , 'ns(YEARS_OF_GIVING_LAST_3)1', 'ns(YEARS_OF_GIVING_LAST_3)2', 'MG_250K_PLUS'
    , 'TRUSTEE_OR_ADVISORY_BOARD'
    , 'Alumnus', 'EVER_PARENT', 'DEEP_ENGAGEMENT', 'CHICAGO_HOME', 'ns(SEASON_TICKET_YEARS)'
    , 'AFFINITY_SCORE', 'MG_PR_MODEL_DESCBottom Tier', 'MG_PR_MODEL_DESCMiddle Tier', 'MG_PR_MODEL_DESCTop Tier'))
) %>% arrange(var)
```
```{r, rows.print = 100}
final_coefs %>% mutate_if(is.numeric, function(x) round(x, 3))
```

  * Unsurprisingly, all intercepts are positive. The campaign all predictors model is closest to 0, and the all predictor models have lower intercepts across the board.
  * Intrestingly, alumni status is negative for all but the largest gift regressed on PG checklist model (which was barely significant).

Let's normalize these using the MG_250K_PLUS indicator appearing in all four models.

```{r, warning = FALSE, message = FALSE, fig.height = 6}
final_coefs %>%
  mutate_if(is.numeric, function(x) {x/x[which(final_coefs$var == 'MG_250K_PLUS')]} %>% round(4)) %>%
  gather('model', 'coefficient', 2:5) %>%
  mutate(model = factor(model, levels = c('campaign.pg', 'largest.pg', 'campaign.all', 'largest.all'))) %>%
  ggplot(aes(x = var, y = coefficient, fill = model)) +
  geom_col(position = position_dodge(), width = .7) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  scale_y_continuous(limits = c(-1, 3), breaks = seq(-1, 3, by = .5)) +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5)
    , axis.title.y = element_text(angle = 0, vjust = .5)
  ) +
  labs(title = 'Model coefficient comparison, scaled to MG = 1', y = bquote(hat(beta)))
```

  * When predicting campaign giving, giving recency (gave in last 3) is more useful than previous MG behavior! Note that loyalty is also a much better predictor than the MG PR model.
  * Visits are also very impactful on average.
  * Most of the coefficients shared between models point in the same direction, with the notable exceptions of age and MG model tiers.

The exact effect of age and recent years of giving are difficult to see since multiple coefficients are computed per spline. Consider the partial residual plots for each model.

```{r}
final_partials <- rbind(
  data.frame(
    model = 'campaign ~ pg score'
    , calc_partial_resids(final_models$campaign.pg.score, inds = c(xval_inds, oos_inds), xname = 'AGE')
  ) %>% mutate(x = mdat %>% filter(rownum %in% unlist(c(xval_inds, oos_inds))) %>% select(NUMERIC_AGE) %>% unlist())
  , data.frame(
      model = 'largest trans ~ pg score'
      , calc_partial_resids(final_models$largest.trans.pg.score, inds = c(xval_inds, oos_inds), xname = 'AGE')
  ) %>% mutate(x = mdat %>% filter(rownum %in% unlist(c(xval_inds, oos_inds))) %>% select(NUMERIC_AGE) %>% unlist())
  , data.frame(
    model = 'campaign ~ all preds'
    , calc_partial_resids(final_models$campaign.all.preds, inds = c(xval_inds, oos_inds)
                          , xname = 'NUMERIC_AGE', df = 5)
  )
  , data.frame(
    model = 'largest trans ~ all preds'
    , calc_partial_resids(final_models$largest.trans.all.preds, inds = c(xval_inds, oos_inds)
                      , xname = 'NUMERIC_AGE', df = 3)
  )
)
```

```{r}
final_partials %>%
  ggplot(aes(x = x, y = y)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 2, color = 'darkgray') +
  geom_line(aes(y = y.hat, color = model), size = 1, alpha = .8) +
  labs(title = 'Numeric age spline versus giving', x = 'age', y = bquote(log[10] ~ 'giving'))
```

```{r}
final_partials %>%
  ggplot(aes(x = x, y = y.partial.resid)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  geom_line(aes(y = y.hat.partial, color = model), size = 1, alpha = .8) +
  labs(title = 'Numeric age spline versus partial residuals for giving'
       , x = 'age'
       , y = 'partial residuals')
```

  * In all four models, age has only a small impact after controlling for the other predictors.
  * The nonlinear splines fits suggest that peak giving occurs in the late 40s or early 50s, and the age 60+ indicator does not provide information beyond that included in other predictors.

```{r}
final_partials <- rbind(
  data.frame(
    model = 'campaign ~ pg score'
    , calc_partial_resids(final_models$campaign.pg.score, inds = c(xval_inds, oos_inds)
                          , xname = 'GAVE_IN_LAST_3_YRS')
  )
  , data.frame(
      model = 'largest trans ~ pg score'
      , calc_partial_resids(final_models$largest.trans.pg.score, inds = c(xval_inds, oos_inds)
                            , xname = 'GAVE_IN_LAST_3_YRS')
  )
  , data.frame(
    model = 'campaign ~ all preds'
    , calc_partial_resids(final_models$campaign.all.preds, inds = c(xval_inds, oos_inds)
                          , xname = 'YEARS_OF_GIVING_LAST_3', df = 2)
  )
  , data.frame(
    model = 'largest trans ~ all preds'
    , calc_partial_resids(final_models$largest.trans.all.preds, inds = c(xval_inds, oos_inds)
                      , xname = 'YEARS_OF_GIVING_LAST_3', df = 2)
  )
)
```

```{r}
final_partials %>%
  ggplot(aes(x = x, y = y)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 2, color = 'darkgray') +
  geom_line(aes(y = y.hat, color = model), size = 1, alpha = .8) +
  labs(title = 'Years of giving last 3 versus giving', x = 'years', y = bquote(log[10] ~ 'giving'))
```

```{r}
final_partials %>%
  ggplot(aes(x = x, y = y.partial.resid)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  geom_line(aes(y = y.hat.partial, color = model, group = model), size = 1, alpha = .8) +
  scale_y_continuous(breaks = -10:10) +
  labs(title = 'Years of giving last 3 versus partial residuals for giving'
       , x = 'years'
       , y = 'partial residuals')
```

```{r}
rbind(
  data.frame(
    model = 'campaign ~ all preds'
    , x = mdat %>% filter(rownum %in% unlist(c(xval_inds, oos_inds))) %>% select(YEARS_OF_GIVING) %>% unlist()
    , y = {final_models$campaign.all.preds %>% model.frame()}[, 1]
    , y.partial.resid = {final_models$campaign.all.preds %>% residuals(type = 'partial')}[, 'sqrt(YEARS_OF_GIVING)']
  ) %>% mutate(y.hat.partial = lm(y.partial.resid ~ sqrt(x)) %>% fitted())
  , data.frame(
    model = 'largest trans ~ all preds'
    , x = mdat %>% filter(rownum %in% unlist(c(xval_inds, oos_inds))) %>% select(YEARS_OF_GIVING) %>% unlist()
    , y = {final_models$largest.trans.all.preds %>% model.frame()}[, 1]
    , y.partial.resid = {final_models$largest.trans.all.preds %>% residuals(type = 'partial')}[, 'sqrt(YEARS_OF_GIVING)']
  ) %>% mutate(y.hat.partial = lm(y.partial.resid ~ sqrt(x)) %>% fitted())
) %>%
  ggplot(aes(x = x, y = y.partial.resid, color = model)) +
  geom_point(alpha = .1, size = 1) +
  geom_hline(yintercept = 0, color = 'darkgray') +
  geom_line(aes(y = y.hat.partial, group = model), size = 1, alpha = .8) +
  scale_y_continuous(breaks = -10:10) +
  labs(title = 'Years of giving versus partial residuals for giving'
       , x = 'years of giving'
       , y = 'partial residuals')
```

  * The PG score models use a binary indicator requiring at least one gift in the last three years, and at least 10 years of giving total, while the all predictors models separate these two components and treat them numerically.
  * Controlling for the effects of the other predictors, recency of giving has a large impact on overall campaign giving in the all predictors model.
  * Looking at years of giving, interestingly both lines cross 0 before 10, suggesting a smaller cutoff (7 or 8 years?) could be used in the PG checklist.
  * When modeling expected campaign giving it's certainly better to treat these variables separately. Largest transaction appears to have a less clear relationship (as seen in the above plots).

# Conclusions

The goal of this project was to look at the impact of each of the PG checklist items has had on giving, and to see whether I could build a reasonable regression model based on the checklist.

The checklist succeeds at its primary goal of identifying factors impacting giving transaction amounts -- in the largest transaction regressed on checklist items model, while alumni status is not significant and president visits are marginal, all the factors are positively associated with giving. A past \$250K+ MG is by far the most important factor, with \$25K+ annual giving, loyalty (gave in last 3 years), and 5+ visits approximately tied for second place. Unfortunately, modeling on largest gift consistently overestimates giving amounts at the high end, and the other variables I examined weren't able to correct for this. The modeled results could still be interesting for prioritization purposes, though.

The big winner was the campaign giving regressed on all predictors model, with nice-looking residual plots exhibiting only a bit of positive skewness. Interestingly, the best predictor by far is donor loyalty, measured by the number of years of giving in the last three, nearly doubling the impact of having made a previous \$250K+ gift!

I also included previous modeled scores (affinity score) and tiering (MG prioritization) in the all predictors models to see whether the PG checklist items add information beyond that contained in these earlier models. Deep engagement, Chicago home address, committee participation, and president visits fall out, but the others are still impactful. Interestingly, alumni status and age don't matter as much as we usually think (alumni status hurts, and age has a nonmonotonic association with giving).

Some ideas for improvements:

  * Evaluation rating significantly impacts giving at some levels but not others, due partially to small sample size but also probably due to binning. Can a continuous rating prediction be developed?

  * Is age a biased predictor? Perhaps age at the time of the largest gift should be used instead of today's age.
  
  * Consider a regression model conditional on making a gift, i.e.
  
$$ E(\text{giving amount} ~ | ~ \text{donor status} = 1) $$
  
  * Consider a time series approach, where future giving over some reasonable cultivation period (2 years?) is estimated conditional on previous data.
  
$$ E \left( (\text{giving amount})_t ~ | ~ (\text{donor attributes})_{t - 1} \right) $$

# Appendix

## Campaign cultivation results {#appendix-campaign-cultivation}

[Back](#campaign-cultivation-score-predictors)

```{r}
lapply(clms, function(x) summary(x))
```

## Campaign all predictors 1 {#appendix-campaign-all-predictors-1}

[Back](#campaign-all-predictors-1)

```{r}
lapply(clmaps, function(x) summary(x))
```

## Campaign all predictors 2 {#appendix-campaign-all-predictors-2}

[Back](#campaign-all-predictors-2)

```{r}
lapply(clmaps2, function(x) summary(x))
```

## Campaign all predictors splines {#appendix-campaign-all-predictors-splines}

[Back](#campaign-all-predictors-splines)

```{r}
lapply(clmaps3
  , function(x) {
    lapply(x, function(y) summary(y))
  }
)
```

## Campaign final models {#appendix-campaign-final-models}

[Back](#campaign-comparison)

```{r}
summary(clm_final)
```
```{r}
summary(clmap_final)
```

## Transaction cultivation results {#appendix-trans-cultivation}

[Back](#trans-cultivation-score-predictors)

```{r}
lapply(tlms, function(x) summary(x))
```

## Transaction all predictors 1 {#appendix-trans-all-1}

[Back](#trans-all-predictors-1)

```{r}
lapply(tlmaps, function(x) summary(x))
```
 
 ## Transaction all predictors 2 {#appendix-trans-all-2}

[Back](#trans-all-predictors-2)

```{r}
lapply(tlmaps2, function(x) summary(x))
```

## Transaction all predictors 2 {#appendix-trans-all-2}

[Back](#trans-all-predictors-2)

```{r}
lapply(tlmaps2, function(x) summary(x))
```

## Transaction final models {#appendix-trans-final-models}

[Back](#trans-comparison)

```{r}
summary(tlm_final)
```
```{r}
summary(tlmap_final)
```

## Final models {#appendix-final-models}

[Back](#final-comparison)

```{r}
lapply(final_models, function(x) summary(x))
```
