==> pg-cultivation-score-fy18/03 Data validation and export.Rmd <==
---
title: "03 Data validation and export"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Goal

Look at descriptive statistics for the final campaign model fitted in the previous [cultivation score weights](https://cdn.rawgit.com/phively/ksm-models/c58e8065/pg-cultivation-score-fy18/02%20Cultivation%20score%20weights.nb.html) step, and calculate fitted values for all KSM entities.

# Setup

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(gridExtra)
library(splines)
library(foreach)

# Functions adapted from previous analysis steps
source('code/functions.R')

# Visualization functions adapted fron previous analysis steps
source('code/functions_viz.R')
```

## Import data

The data file is generated with [this code](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/code/generate-data.R), and the modeled data frame with [this code](https://github.com/phively/ksm-models/blob/master/pg-cultivation-score-fy18/code/generate-modeling-data.R).

```{r}
filepath <- 'data/2018-07-26 PG scores for all active prospects.xlsx'
source('code/generate-data.R')
source('code/generate-modeling-data.R')
```

## Build model

Re-fit the [all predictors 2](https://cdn.rawgit.com/phively/ksm-models/c58e8065/pg-cultivation-score-fy18/02%20Cultivation%20score%20weights.nb.html#all_predictors_2) model.

```{r}
clm_alldat <- lm(
  log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
    ns(NUMERIC_AGE, df = 5) + # Underlying variable to AGE indicator
    PM_VISIT_LAST_2_YRS +
    log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
    AF_25K_GIFT +
    sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
    ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
    MG_250K_PLUS + # Decided to leave as factor
    Alumnus +
    SEASON_TICKET_YEARS + # Deep Engagement component
    AFFINITY_SCORE +
    MG_PR_MODEL_DESC
  , data = mdat
)
```

```{r}
summary(clm_alldat)
```

Residuals plot recreated in ggplot2.

```{r}
plot_resids(list(clm_alldat), list(1:nrow(mdat), 1:nrow(mdat)), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$insample +
  geom_point(alpha = .05) +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  guides(color = FALSE)
```

Q-Q plot recreated in ggplot2.

```{r}
plot_qq(list(clm_alldat), xval = list(1:nrow(mdat), 1:nrow(mdat)), yname= 'CAMPAIGN_NEWGIFT_CMIT_CREDIT')$insample +
  scale_y_continuous(breaks = seq(-10, 10, by = 2)) +
  guides(color = FALSE)
```

# Descriptives

How well-calibrated is the model at different levels? Look at the true giving amounts versus predicted ones.

```{r}
# Predictions data frame
model_preds <- calc_preds(list(clm_alldat), list(1:nrow(mdat)), 'CAMPAIGN_NEWGIFT_CMIT_CREDIT') %>%
  data.frame() %>%
  mutate(
    resid = truth - preds
    , cgift = log10plus1(truth, inverse = TRUE)
    , cgift.preds = log10plus1(preds, inverse = TRUE)
    , cgift.bin = trunc(truth)
    , cgift.preds.bin = trunc(preds)
  )
```
```{r}
# Plot predictions against truth
model_preds %>%
  ggplot(aes(x = cgift, y = cgift.preds)) +
  geom_abline(color = 'purple', alpha = .5) +
  geom_point(alpha = .05) +
  geom_smooth(color = 'red') +
  geom_smooth(method = 'lm') +
  scale_x_continuous(trans = 'log10plus1', breaks = 10^(0:10), minor_breaks = NULL, labels = scales::dollar) +
  scale_y_continuous(trans = 'log10plus1', breaks = 10^(0:10), minor_breaks = NULL, labels = scales::dollar) +
  coord_fixed() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Campaign giving', y = 'Predicted campaign giving')
```

```{r}
mingift <- -1
maxgift <- 9
model_preds %>%
  group_by(cgift.bin, cgift.preds.bin) %>%
  summarise(n = length(resid)) %>%
  ggplot(aes(x = cgift.bin, y = cgift.preds.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Campaign giving', y = 'Predicted campaign giving')
```

Percent of donors predicted within one power of ten of their correct giving level:

```{r}
preds.table <- model_preds %>%
  arrange(cgift.bin) %>%
  mutate(
    call = case_when(
      resid >= 1 ~ 'Underestimated' # Understimated by more than power of 10
      , resid <= -1 ~ 'Overestimated'  # Overestimated by more than power of 10
      , TRUE ~ 'Within power of 10' # Correct estiamte
    ) %>% as_factor(levels = c('Within power of 10', 'Underestimated', 'Overestimated'))
    , `Campaign Giving Range` = 10^cgift.bin %>% scales::dollar() %>% as_factor()
  ) %>%
  group_by(`Campaign Giving Range`, call) %>%
  summarise(n = length(resid)) %>%
  spread(call, n, fill = 0) %>%
  mutate(
    Total = Overestimated + `Within power of 10` + Underestimated
    , `Over%` = {Overestimated / Total} %>% scales::percent()
    , `Within%` = {`Within power of 10` / Total}  %>% scales::percent()
    , `Under%` = {Underestimated / Total}  %>% scales::percent()
  )
```

```{r, rows.print = 100}
print(preds.table)
```

I'd interpret the group of people who actually gave \$10K or \$100K and were estimated at \$1M+ as prospets, and the people who actually gave \$1M+ but were estimated below \$100K could be looked into to search for additional possible factors.

# Append scores

```{r}
pool %>%
  mutate(
    rownum = 1:nrow(pool), actual = log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT)
    , fitted = lm(
        log10plus1(CAMPAIGN_NEWGIFT_CMIT_CREDIT) ~
          ns(NUMERIC_AGE, df = 5) + # Underlying variable to AGE indicator
          PM_VISIT_LAST_2_YRS +
          log10plus1(VISIT_COUNT) + # Underlying VISITS_5PLUS indicator
          AF_25K_GIFT +
          sqrt(YEARS_OF_GIVING) + # Underlying GAVE_IN_LAST_3_YRS
          ns(YEARS_OF_GIVING_LAST_3, df = 2) + # Underlying GAVE_IN_LAST_3_YRS
          MG_250K_PLUS + # Decided to leave as factor
          Alumnus +
          SEASON_TICKET_YEARS + # Deep Engagement component
          AFFINITY_SCORE +
          MG_PR_MODEL_DESC
        , data = mdat
      ) %>% fitted()
    , resid = actual - fitted
    , underestimated_pg = case_when(resid >= 1 & trunc(actual) >= 6 ~ 'Y')
  ) %>%
  write_csv(path = 'data/fitted.csv')
```
