==> pg-cultivation-score-fy18/05b GAM experimentation.Rmd <==
---
title: "05b GAM experimentation"
output:
  html_document:
    df_print: paged
    toc: yes
  html_notebook:
    code_folding: hide
    toc: yes
    toc_float:
      collapsed: no
---

# Goal

I'm curious about how generalized additive models perform relative to the less involved methods

# Setup

```{r setup, message = FALSE, warning = FALSE}
library(tidyverse)
library(reshape2)
library(gridExtra)
library(splines)
library(lubridate)
library(wranglR)
library(Boruta)
library(foreach)
library(doParallel)
library(glmnet)
library(glmnetUtils)
library(mgcv)

# Functions adapted from previous analysis steps
source('code/functions.R')

# Visualization functions adapted fron previous analysis steps
source('code/functions_viz.R')

# Set number of available CPU cores
registerDoParallel(detectCores() - 2)
```
```{r}
# Parameters
train_fy <- 2016
filepath <- 'data/2018-11-30 point-in-time data.xlsx'
sheetname <- 'Select point_in_time_model'

# Import data
source('code/generate-pit-data.R')

# Run data generation functions
modeling.data <- generate_pit_data(filepath, sheetname) %>%
  generate_additional_predictors()
```

# Generalized additive models

Fit a few different models for comparison.

```{r}
set.seed(123)
mdat <- KFoldXVal(modeling.data, k = 2, prop = .8)
train <- modeling.data[mdat[[1]], ]
test <- modeling.data[mdat[[2]], ]
```

```{r}
gmod <- train %>%
  gam(
  rv.amt ~ COMMITTEE_KSM_LDR + s(CRU_PFY1) + s(CRU_PFY2) +
    s(CRU_PFY3) + s(CRU_PFY4) + s(CRU_PFY5) + s(EVALUATION_LOWER_BOUND) +
    GIFT_CLUB_NU_LDR_YRS + s(GIFTS_ALLOCS_SUPPORTED) + s(GIFTS_OUTRIGHTS_PAYMENTS) +
    s(GIVING_CRU_TOTAL) + s(GIVING_MAX_CASH_YR) + s(GIVING_MAX_PLEDGE_AMT) +
    s(GIVING_MAX_PLEDGE_FY) + HOUSEHOLD_CONTINENT + KSM_GOS_FLAG + LOYAL_5_PCT_ANY + s(NGC_PFY1) + 
    s(NGC_PFY2) + s(NGC_PFY3) + s(NGC_PFY4) + s(NGC_PFY5) + PREF_ADDR_TYPE_CODE + 
    PROGRAM_GROUP + s(RECORD_YR) + s(UOR_LOWER_BOUND) + UPGRADE3_CASH + s(VELOCITY3_LIN_NGC) + 
    s(VISITS_5FY)
  , data = .
  , family = 'gaussian'
)
```
```{r}
lmod <- train %>%
  lm(
  rv.amt ~ COMMITTEE_KSM_LDR + ns(CRU_PFY1, df = 4) + 
    ns(CRU_PFY2, df = 4) + ns(CRU_PFY3, df = 4) + ns(CRU_PFY4, 
    df = 4) + ns(CRU_PFY5, df = 4) + ns(EVALUATION_LOWER_BOUND, 
    df = 4) + GIFT_CLUB_NU_LDR_YRS + ns(GIFTS_ALLOCS_SUPPORTED, 
    df = 4) + ns(GIFTS_OUTRIGHTS_PAYMENTS, df = 4) + ns(GIVING_CRU_TOTAL, 
    df = 4) + ns(GIVING_MAX_CASH_YR, df = 4) + ns(GIVING_MAX_PLEDGE_AMT, 
    df = 4) + ns(GIVING_MAX_PLEDGE_FY, df = 4) + HOUSEHOLD_CONTINENT + 
    KSM_GOS_FLAG + LOYAL_5_PCT_ANY + ns(NGC_PFY1, df = 4) + 
    ns(NGC_PFY2, df = 4) + ns(NGC_PFY3, df = 4) + ns(NGC_PFY4, 
    df = 4) + ns(NGC_PFY5, df = 4) + PREF_ADDR_TYPE_CODE + 
    PROGRAM_GROUP + ns(RECORD_YR, df = 4) + ns(UOR_LOWER_BOUND, 
    df = 4) + UPGRADE3_CASH + ns(VELOCITY3_LIN_NGC, df = 4) + 
    ns(VISITS_5FY, df = 4)
  , data = .
)
```
```{r}
lmod2 <- train %>% filter(rv.gave) %>%
  lm(
  rv.amt ~ COMMITTEE_KSM_LDR + ns(CRU_PFY1, df = 4) + 
    ns(CRU_PFY2, df = 4) + ns(CRU_PFY3, df = 4) + ns(CRU_PFY4, 
    df = 4) + ns(CRU_PFY5, df = 4) + ns(EVALUATION_LOWER_BOUND, 
    df = 4) + GIFT_CLUB_NU_LDR_YRS + ns(GIFTS_ALLOCS_SUPPORTED, 
    df = 4) + ns(GIFTS_OUTRIGHTS_PAYMENTS, df = 4) + ns(GIVING_CRU_TOTAL, 
    df = 4) + ns(GIVING_MAX_CASH_YR, df = 4) + ns(GIVING_MAX_PLEDGE_AMT, 
    df = 4) + ns(GIVING_MAX_PLEDGE_FY, df = 4) + HOUSEHOLD_CONTINENT + 
    KSM_GOS_FLAG + LOYAL_5_PCT_ANY + ns(NGC_PFY1, df = 4) + 
    ns(NGC_PFY2, df = 4) + ns(NGC_PFY3, df = 4) + ns(NGC_PFY4, 
    df = 4) + ns(NGC_PFY5, df = 4) + PREF_ADDR_TYPE_CODE + 
    PROGRAM_GROUP + ns(RECORD_YR, df = 4) + ns(UOR_LOWER_BOUND, 
    df = 4) + UPGRADE3_CASH + ns(VELOCITY3_LIN_NGC, df = 4) + 
    ns(VISITS_5FY, df = 4)
  , data = .
)
```
```{r}
glmod <- train %>%
  glm(
  rv.gave ~ PROGRAM_GROUP + PREF_ADDR_TYPE_CODE + 
    HOUSEHOLD_CONTINENT + BUS_IS_EMPLOYED + HAS_HOME_ADDR + HAS_HOME_PHONE + 
    ns(YEARS_SINCE_FIRST_GIFT, df = 4) + ns(GIVING_FIRST_YEAR_CASH_AMT, 
    df = 4) + ns(GIVING_MAX_PLEDGE_AMT, df = 4) + ns(GIVING_CASH_TOTAL, 
    df = 4) + ns(GIVING_PLEDGE_TOTAL, df = 4) + ns(GIVING_CRU_TOTAL, 
    df = 4) + ns(GIFTS_ALLOCS_SUPPORTED, df = 4) + ns(GIFTS_FYS_SUPPORTED, 
    df = 4) + ns(GIFTS_CASH, df = 4) + ns(GIFTS_PLEDGES, 
    df = 4) + ns(CASH_PFY1, df = 4) + ns(CASH_PFY2, df = 4) + 
    ns(CASH_PFY3, df = 4) + ns(CASH_PFY4, df = 4) + ns(CASH_PFY5, 
    df = 4) + CRU_GIVING_SEGMENT + ns(EVALUATION_LOWER_BOUND, 
    df = 4) + ns(UOR_LOWER_BOUND, df = 4) + ns(MONTHS_ASSIGNED, 
    df = 4) + ns(COMMITTEE_NU_DISTINCT, df = 4) + ns(COMMITTEE_NU_YEARS, 
    df = 4) + ns(COMMITTEE_KSM_DISTINCT, df = 4) + ns(EVENTS_PREV_3_FY, 
    df = 4) + ns(EVENTS_CFY, df = 4) + ns(EVENTS_PFY1, df = 4) + 
    ns(ATHLETICS_TICKET_YEARS, df = 4) + ns(YEARS_SINCE_ATHLETICS_TICKETS, 
    df = 4) + ns(RECORD_YR, df = 4) + ns(YEARS_SINCE_MAX_CASH_YR, 
    df = 4) + GIVING_MAX_CASH_MO + KSM_PROSPECT + ns(VISITORS_5FY, 
    df = 4) + LOYAL_5_PCT_CASH + UPGRADE3_CASH + VELOCITY3_LIN_CASH + 
    SPOUSE_ALUM
  , data = .
  , family = 'binomial'
)
```
```{r}
gamp <- train %>%
  gam(
  rv.gave ~ PROGRAM_GROUP + PREF_ADDR_TYPE_CODE + 
    HOUSEHOLD_CONTINENT + BUS_IS_EMPLOYED + HAS_HOME_ADDR + HAS_HOME_PHONE + 
    s(YEARS_SINCE_FIRST_GIFT) + s(GIVING_FIRST_YEAR_CASH_AMT)
 + s(GIVING_MAX_PLEDGE_AMT) + s(GIVING_CASH_TOTAL)
 + s(GIVING_PLEDGE_TOTAL) + s(GIVING_CRU_TOTAL)
 + s(GIFTS_ALLOCS_SUPPORTED) + s(GIFTS_FYS_SUPPORTED)
 + s(GIFTS_CASH) + s(GIFTS_PLEDGES, k = 4)
 + s(CASH_PFY1) + s(CASH_PFY2) + 
    s(CASH_PFY3) + s(CASH_PFY4) + s(CASH_PFY5)
 + CRU_GIVING_SEGMENT + s(EVALUATION_LOWER_BOUND)
 + s(UOR_LOWER_BOUND) + s(MONTHS_ASSIGNED)
 + s(COMMITTEE_NU_DISTINCT) + s(COMMITTEE_NU_YEARS)
 + s(COMMITTEE_KSM_DISTINCT) + s(EVENTS_PREV_3_FY)
 + s(EVENTS_CFY) + s(EVENTS_PFY1) + 
    s(ATHLETICS_TICKET_YEARS) + s(YEARS_SINCE_ATHLETICS_TICKETS)
 + s(RECORD_YR) + s(YEARS_SINCE_MAX_CASH_YR)
 + GIVING_MAX_CASH_MO + KSM_PROSPECT + s(VISITORS_5FY)
 + LOYAL_5_PCT_CASH + UPGRADE3_CASH + VELOCITY3_LIN_CASH + 
    SPOUSE_ALUM
  , data = .
  , family = 'binomial'
)
```
```{r}
game <- train %>% filter(rv.gave) %>%
  gam(
  rv.amt ~ COMMITTEE_KSM_LDR + s(CRU_PFY1) + s(CRU_PFY2) +
    s(CRU_PFY3) + s(CRU_PFY4) + s(CRU_PFY5) + s(EVALUATION_LOWER_BOUND) +
    GIFT_CLUB_NU_LDR_YRS + s(GIFTS_ALLOCS_SUPPORTED) + s(GIFTS_OUTRIGHTS_PAYMENTS) +
    s(GIVING_CRU_TOTAL) + s(GIVING_MAX_CASH_YR) + s(GIVING_MAX_PLEDGE_AMT) +
    s(GIVING_MAX_PLEDGE_FY) + HOUSEHOLD_CONTINENT + KSM_GOS_FLAG + LOYAL_5_PCT_ANY + s(NGC_PFY1) + 
    s(NGC_PFY2) + s(NGC_PFY3) + s(NGC_PFY4) + s(NGC_PFY5) + PREF_ADDR_TYPE_CODE + 
    PROGRAM_GROUP + s(RECORD_YR) + s(UOR_LOWER_BOUND) + UPGRADE3_CASH + s(VELOCITY3_LIN_NGC) + 
    s(VISITS_5FY)
  , data = .
  , family = 'gaussian'
  , control = list(nthreads = 10)
)
```

```{r}
summary(gmod)
```

```{r}
summary(lmod)
```

```{r}
preds <- data.frame(
  giving = test$rv.amt
  , gam = predict(gmod, newdata = test)
  , lm = predict(lmod, newdata = test)
  , gam_part = predict(game, newdata = test)
  , lm_part = predict(lmod2, newdata = test)
  , glm = predict(glmod, newdata = test, type = 'response')
  , gamp = predict(gamp, newdata = test, type = 'response')
) %>% mutate(
  giving_bin = trunc(giving)
  , gam_bin = trunc(gam)
  , lm_bin = trunc(lm)
  , gam_part_bin = trunc(gam_part)
  , lm_part_bin = trunc(lm_part)
  , split_bin = trunc(lm_part * glm)
  , dbl_gam_bin = trunc(gam_part * gamp)
)
```
```{r}
data.frame(
  gam_alldat = calc_mse(y = test$rv.amt, yhat = preds$gam)
  , lm_alldat = calc_mse(y = test$rv.amt, yhat = preds$lm)
  , gam_part = calc_mse(y = test$rv.amt, yhat = preds$gam_part)
  , lm_part = calc_mse(y = test$rv.amt, yhat = preds$lm_part)
  , split_model = calc_mse(y = test$rv.amt, yhat = preds$glm * preds$lm_part)
  , split_gam = calc_mse(y = test$rv.amt, yhat = preds$gam_part * preds$glm)
  , dbl_gam = calc_mse(y = test$rv.amt, yhat = preds$gam_part * preds$gamp)
)
```

```{r}
mingift <- -1
maxgift <- 8
a <- preds %>% 
  ggplot(aes(x = giving, y = gam)) +
  geom_point(color = 'gray', alpha = .5) +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
b <- preds %>% 
  ggplot(aes(x = giving, y = lm)) +
  geom_point(color = 'gray', alpha = .5) +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
c <- preds %>%
  mutate(split = lm_part * glm) %>%
  ggplot(aes(x = giving, y = split)) +
  geom_point(color = 'gray', alpha = .5) +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
d <- preds %>%
  mutate(dgam = gam_part * gamp) %>%
  ggplot(aes(x = giving, y = dgam)) +
  geom_point(color = 'gray', alpha = .5) +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
```
```{r, fig.width = 10, fig.height = 10}
grid.arrange(a, b, c, d)
```

```{r}
a <- preds %>% 
  group_by(giving_bin, gam_bin) %>%
  summarise(n = length(giving), g = sum(giving)) %>%
    ggplot(aes(x = giving_bin, y = gam_bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
b <- preds %>% 
  group_by(giving_bin, lm_bin) %>%
  summarise(n = length(giving), g = sum(giving)) %>%
    ggplot(aes(x = giving_bin, y = lm_bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
c <- preds %>% 
  group_by(giving_bin, split_bin) %>%
  summarise(n = length(giving), g = sum(giving)) %>%
    ggplot(aes(x = giving_bin, y = split_bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
d <- preds %>% 
  group_by(giving_bin, dbl_gam_bin) %>%
  summarise(n = length(giving), g = sum(giving)) %>%
    ggplot(aes(x = giving_bin, y = dbl_gam_bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5))
```
```{r, fig.width = 10, fig.height = 10}
grid.arrange(a, b, c, d)
```

Surprisingly to me, the lm and gam performance is very close.
