==> pg-cultivation-score-fy18/11 FY19 refresh.Rmd <==
---
title: "11 FY19 refresh"
output:
  html_notebook:
    toc: TRUE
    toc_float:
      collapsed: FALSE
---

# Setup

```{r setup, message = FALSE, warning = FALSE}
# Document parameters
# Look for # <UPDATE> tags
threads <- 3 # <UPDATE>

# Packages
library(tidyverse)
library(reshape2)
library(gridExtra)
library(splines)
library(lubridate)
library(wranglR)
library(Boruta)
library(foreach)
library(doParallel)
library(mgcv)

# Functions adapted from previous analysis steps
source('code/functions.R')

# Visualization functions adapted fron previous analysis steps
source('code/functions_viz.R')

# Set number of available CPU cores
registerDoParallel(threads)
```

```{r}
# Load current data
# Import data
train_fy <- 2017 # <UPDATE>
source('code/generate-pit-data.R')

# Run data generation functions
# Ensure that the file pointed to has the correct train_fy (2 FY ago)
full.data <- generate_pit_data(
  filepath = 'data/2019-10-04 point-in-time data.xlsx' # <UPDATE>
  , sheetname = 'Select point_in_time_model_2019' # <UPDATE>
)
```

```{r}
# Updated dataset
modeling.data <- full.data %>%
  generate_additional_predictors()
```

# Updated models

I'm reusing the model parameters from FY18-19; see [06 KSM model tuning](https://phively.github.io/ksm-models/pg-cultivation-score-fy18/06%20KSM%20model%20tuning.nb.html#probability_model_tuning)

```{r}
# Model parameters
giving <- 3
dollars <- 2
engagement <- 3
recordyr <- 5
```

Updating the GAM (unconditional probability).

```{r}
gam_final <- gam(
  rv.gave ~
  PROGRAM_GROUP +
  PREF_ADDR_TYPE_CODE +
  HOUSEHOLD_CONTINENT +
  BUS_IS_EMPLOYED +
  HAS_HOME_ADDR +
  HAS_HOME_PHONE +
  s(YEARS_SINCE_FIRST_GIFT, k = giving) +
  s(GIVING_FIRST_YEAR_CASH_AMT, k = giving) +
  s(GIVING_MAX_PLEDGE_AMT, k = giving) +
  s(GIVING_CASH_TOTAL, k = giving) +
  s(GIVING_PLEDGE_TOTAL, k = giving) +
  s(GIVING_CRU_TOTAL, k = giving) +
  s(GIFTS_ALLOCS_SUPPORTED, k = giving) +
  s(GIFTS_FYS_SUPPORTED, k = giving) +
  s(GIFTS_CASH, k = giving) +
  s(GIFTS_PLEDGES, k = giving) +
  s(CASH_PFY1, k = dollars) +
  s(CASH_PFY2, k = dollars) +
  s(CASH_PFY3, k = dollars) +
  s(CASH_PFY4, k = dollars) +
  s(CASH_PFY5, k = dollars) +
  CRU_GIVING_SEGMENT +
  s(EVALUATION_LOWER_BOUND, k = dollars) +
  s(UOR_LOWER_BOUND, k = dollars) +
  s(MONTHS_ASSIGNED, k = engagement) +
  s(COMMITTEE_NU_DISTINCT, k = engagement) +
  s(COMMITTEE_NU_YEARS, k = engagement) +
  s(COMMITTEE_KSM_DISTINCT, k = engagement) +
  s(EVENTS_PREV_3_FY, k = engagement) +
  s(EVENTS_CFY, k = engagement) +
  s(EVENTS_PFY1, k = engagement) +
  s(ATHLETICS_TICKET_YEARS, k = engagement) +
  s(YEARS_SINCE_ATHLETICS_TICKETS, k = engagement) +
  s(RECORD_YR, k = recordyr) +
  s(YEARS_SINCE_MAX_CASH_YR, k = giving) +
  GIVING_MAX_CASH_MO +
  KSM_PROSPECT +
  s(VISITORS_5FY, k = engagement) +
  LOYAL_5_PCT_CASH +
  UPGRADE3_CASH +
  VELOCITY3_LIN_CASH +
  SPOUSE_ALUM
  , data = modeling.data
  , family = 'binomial'
  , control = list(nthreads = threads)
)
```

Updating the LM (conditional expectation).

```{r}
lm_final <- lm(
  rv.amt ~
  COMMITTEE_KSM_LDR +
  ns(CRU_PFY1, df = 2) +
  ns(CRU_PFY2, df = 1) +
  ns(CRU_PFY3, df = 1) +
  ns(CRU_PFY4, df = 1) +
  ns(CRU_PFY5, df = 1) +
  ns(EVALUATION_LOWER_BOUND, df = 1) +
  GIFT_CLUB_NU_LDR_YRS +
  ns(GIFTS_ALLOCS_SUPPORTED, df = 1) +
  ns(GIFTS_OUTRIGHTS_PAYMENTS, df = 1) +
  ns(GIVING_CRU_TOTAL, df = 1) +
  ns(GIVING_MAX_CASH_YR, df = 1) +
  ns(GIVING_MAX_PLEDGE_AMT, df = 1) +
  ns(GIVING_MAX_PLEDGE_FY, df = 1) +
  HOUSEHOLD_CONTINENT +
  KSM_GOS_FLAG +
  LOYAL_5_PCT_ANY +
  ns(NGC_PFY1, df = 2) +
  ns(NGC_PFY2, df = 1) +
  ns(NGC_PFY3, df = 1) +
  ns(NGC_PFY4, df = 1) +
  ns(NGC_PFY5, df = 1) +
  PREF_ADDR_TYPE_CODE +
  PROGRAM_GROUP +
  ns(RECORD_YR, df = 5) +
  ns(UOR_LOWER_BOUND, df = 1) +
  UPGRADE3_CASH +
  ns(VELOCITY3_LIN_NGC, df = 7) +
  ns(VISITS_5FY, df = 1)
  , data = modeling.data %>% filter(rv.gave == TRUE)
)
```

# Generate new predictions

```{r}
# Load current data
# Parameters
train_fy <- 2019 # <UPDATE>
filepath <- 'data/2019-10-08 point-in-time fy19 future data.xlsx' # <UPDATE>
sheetname <- 'Select point_in_time_model_19fu' # <UPDATE>

# Run data generation functions
# Ensure that the file pointed to has the correct train_fy (current FY)
new.full.data <- generate_pit_data(filepath, sheetname)
new.modeling.data <- new.full.data %>% generate_additional_predictors(future.data = TRUE)
```


```{r}
preds_prb <- predict(gam_final, newdata = new.modeling.data, type = 'response') %>% as.numeric()
preds_amt <- predict(lm_final, newdata = new.modeling.data, type = 'response')
```

# New prediction diagnostics

```{r}
data.frame(
  data = 'insample'
  , prediction = gam_final$fitted.values
) %>% rbind(
  data.frame(
    data = 'outsample'
    , prediction = preds_prb
  )
) %>%
  ggplot(aes(x = prediction, fill = data)) +
  geom_histogram(bins = 20) +
  facet_wrap(. ~ data) +
  labs(title = 'Probability model fitted values')
```

Looks fine.

```{r}
data.frame(
  data = 'insample'
  , prediction = predict(lm_final, newdata = modeling.data)
) %>% rbind(
  data.frame(
    data = 'outsample'
    , prediction = preds_amt
  )
) %>%
  ggplot(aes(x = prediction, fill = data)) +
  geom_histogram(bins = 20) +
  facet_wrap(. ~ data) +
  labs(title = 'Gift amount model fitted values')
```

Looks fine.

```{r}
data.frame(
  id = modeling.data$ID_NUMBER %>% as.numeric()
  , p_insample = gam_final$fitted.values
) %>% inner_join(
  data.frame(
    id = new.modeling.data$ID_NUMBER %>% as.numeric()
    , p_outsample = preds_prb
  )
  , by = c('id')
) %>% mutate(
  change = p_insample - p_outsample
) %>%
  ggplot(aes(x = change)) + 
  geom_histogram(bins = 20, alpha = .5) +
  scale_x_continuous(breaks = seq(-1, 1, by = .2)) +
  scale_y_log10(breaks = 10^(0:6), labels = format(10^(0:6), scientific = FALSE)) +
  labs(title = 'Change in predicted probability of giving')
```

Some pretty large swings but very few over $\pm .5$ (log scale) and the near-symmetry is reassuring.


```{r}
data.frame(
  id = modeling.data$ID_NUMBER %>% as.numeric()
  , p_insample = predict(lm_final, newdata = modeling.data)
) %>% inner_join(
  data.frame(
    id = new.modeling.data$ID_NUMBER %>% as.numeric()
    , p_outsample = preds_amt
  )
  , by = c('id')
) %>% mutate(
  change = p_insample - p_outsample
) %>%
  ggplot(aes(x = change)) + 
  geom_histogram(bins = 16, alpha = .5) +
  scale_y_log10(breaks = 10^(0:6), labels = format(10^(0:6), scientific = FALSE)) +
  labs(title = 'Change in predicted giving amounts')
```

Looks fine, and the near-symmetry is reassuring.

# Previous model comparison

```{r}
fy18_prb <- predict(gam_final_18, newdata = new.modeling.data, type = 'response') %>% as.numeric()
fy18_amt <- predict(lm_final_18, newdata = new.modeling.data, type = 'response')
```

```{r}
# FY18 data max EV score (PR)
idx <- which(new.modeling.data$ID_NUMBER == '0000372980')
data.frame(
  p18 = fy18_prb[idx]
  , p19 = preds_prb[idx]
  , g18 = fy18_amt[idx]
  , g19 = preds_amt[idx]
) %>% mutate(
  ev18 = p18 * g18
  , ev19 = p19 * g19
)
```

Interestingly, the amount models are pretty close but the probabilities are very different.

## LM comparison

```{r, rows.print = 100}
coef18_lm <- summary(lm_final_18)$coef %>% data.frame() %>%
  select(Estimate18 = Estimate, SE18 = Std..Error, Pr.t18 = Pr...t..)
coef19_lm <- summary(lm_final)$coef %>% data.frame() %>%
  select(Estimate19 = Estimate, SE19 = Std..Error, Pr.t19 = Pr...t..)

cbind(
  coef18_lm
  , coef19_lm
) %>% mutate(
  varname = rownames(coef18_lm)
  , delta = Estimate19 - Estimate18
  , delta_pct = {delta / abs(Estimate18)} %>% scales::percent()
) %>% select(
  varname
  , Estimate18
  , Estimate19
  , delta
  , delta_pct
  , SE18
  , SE19
  , Pr.t18
  , Pr.t19
) %>% mutate_if(
  is.numeric, signif, 3
)
```

## GAM comparison

```{r, rows.print = 50}
coef18_gam <- summary(gam_final_18)
coef19_gam <- summary(gam_final)

data.frame(
  names = coef18_gam$p.coeff %>% names()
  , coef18 = coef18_gam$p.coeff
  , coef19 = coef19_gam$p.coeff
) %>% mutate(
  delta = coef19 - coef18
  , delta_pct = {delta / abs(coef18)} %>% scales::percent()
  , t18 = coef18_gam$p.t
  , t19 = coef19_gam$p.t
)
```

There are some very wide swings in how certain variables, such as upgrade amounts, are treated. I'm thinking these should be handled differently in the future when I refit a new model.

# Diagnostics

```{r}
# Diagnostic data
error_compare <- data.frame(
  giving = log10plus1(modeling.data$rv.amt, inverse = TRUE)
  , log.giving = modeling.data$rv.amt
  , gave = modeling.data$rv.gave
  , est.giving = predict(lm_final, newdata = modeling.data, type = 'response')
  , est.gave = gam_final$fitted
) %>% mutate(
  ev = est.giving * est.gave
  , error = log.giving - ev
  , log.giving.bin = trunc(log.giving)
  , pred.giving.bin = trunc(est.giving)
  , ev.bin = trunc(ev)
)

error_compare_18 <- data.frame(
  giving = log10plus1(modeling.data$rv.amt, inverse = TRUE)
  , log.giving = modeling.data$rv.amt
  , gave = modeling.data$rv.gave
  , est.giving = predict(lm_final_18, newdata = modeling.data, type = 'response')
  , est.gave = predict(gam_final_18, newdata = modeling.data, type = 'response') %>% as.numeric()
) %>% mutate(
  ev = est.giving * est.gave
  , error = log.giving - ev
  , log.giving.bin = trunc(log.giving)
  , pred.giving.bin = trunc(est.giving)
  , ev.bin = trunc(ev)
)
```

```{r, fig.width = 8}
grid.arrange(
error_compare_18 %>%
  ggplot(aes(x = log.giving, y = est.giving, color = gave)) +
  geom_point() +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  labs(title = 'coef1 estimated versus actual giving')
, error_compare %>%
  ggplot(aes(x = log.giving, y = est.giving, color = gave)) +
  geom_point() +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  labs(title = 'coef2 estimated versus actual giving')
, ncol = 2
)
```

```{r}
data.frame(
  model = c('coef1', 'coef2')
  , mse = c(
    calc_mse(y = error_compare$log.giving, yhat = error_compare$est.giving)
    , calc_mse(y = error_compare_18$log.giving, yhat = error_compare_18$est.giving)
  )
)
```

```{r, fig.width = 8}
grid.arrange(
error_compare_18 %>%
  ggplot(aes(x = log.giving, y = ev, color = gave)) +
  geom_point() +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  labs(title = 'coef1 expected value versus actual giving')
, error_compare %>%
  ggplot(aes(x = log.giving, y = ev, color = gave)) +
  geom_point() +
  geom_abline(aes(slope = 1, intercept = 0)) +
  geom_smooth(method = 'gam', formula = y ~ s(x, bs = 'cs'), color = 'blue') +
  labs(title = 'coef2 expected value versus actual giving')
, ncol = 2
)
```

```{r}
data.frame(
  model = c('coef1', 'coef2')
  , mse = c(
    calc_mse(y = error_compare_18$log.giving, yhat = error_compare_18$ev)
    , calc_mse(y = error_compare$log.giving, yhat = error_compare$ev)
  )
)
```

The coef2 model trend line is non-monotonic which is pretty odd.

```{r, fig.width = 6, fig.height = 8}
error_bin <- error_compare %>%
  group_by(log.giving.bin, pred.giving.bin) %>%
  summarise(n = length(gave), g = sum(gave))

error_bin_18 <- error_compare_18 %>%
  group_by(log.giving.bin, pred.giving.bin) %>%
  summarise(n = length(gave), g = sum(gave))

mingift <- -1
maxgift <- 8

grid.arrange(
error_bin_18 %>%
  ggplot(aes(x = log.giving.bin, y = pred.giving.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Actual giving', y = 'coef1 Predicted giving')
, error_bin %>%
  ggplot(aes(x = log.giving.bin, y = pred.giving.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Actual giving', y = 'coef2 Predicted giving')
)
```

__coef1 model__

  * Total giving during the out-of-sample time period was `r tmpd <- error_compare_18%>% select(giving) %>% sum(); tmpd %>% scales::dollar() %>% I()`.
  * Those predicted to give at least `r tmpcutoff <- 3; scales::dollar(10^tmpcutoff) %>% I()` gave `r tmpn <- error_compare_18%>% filter(pred.giving.bin >= tmpcutoff) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or `r (tmpn/tmpd) %>% scales::percent() %>% I()` of the total.
  * A total of `r tmpd <- error_compare_18%>% filter(pred.giving.bin >= tmpcutoff) %>% nrow(); tmpd %>% I()` donors were predicted to give at the `r scales::dollar(10^tmpcutoff) %>% I()` level and `r tmpn <- error_compare_18%>% filter(pred.giving.bin >= tmpcutoff & log.giving.bin >= tmpcutoff) %>% nrow(); tmpn %>% I()` actually did, for a hit rate of `r (tmpn/tmpd) %>% scales::percent() %>% I()`.
  * Average giving in this group is `r ({error_compare_18%>% filter(pred.giving.bin >= tmpcutoff) %>% select(giving) %>% sum()}/tmpd) %>% scales::dollar() %>% I()` per person.
  * This method elimintates `r tmpd <- error_compare_18%>% filter(pred.giving.bin < tmpcutoff) %>% select(giving) %>% nrow(); tmpd %>% I()` entities, who gave a total of `r tmpn <- error_compare_18%>% filter(pred.giving.bin < tmpcutoff) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or an average of `r (tmpn/tmpd) %>% scales::dollar() %>% I()` per person.

__coef2 model__

  * Total giving during the out-of-sample time period was `r tmpd <- error_compare %>% select(giving) %>% sum(); tmpd %>% scales::dollar() %>% I()`.
  * Those predicted to give at least `r tmpcutoff <- 3; scales::dollar(10^tmpcutoff) %>% I()` gave `r tmpn <- error_compare %>% filter(pred.giving.bin >= tmpcutoff) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or `r (tmpn/tmpd) %>% scales::percent() %>% I()` of the total.
  * A total of `r tmpd <- error_compare %>% filter(pred.giving.bin >= tmpcutoff) %>% nrow(); tmpd %>% I()` donors were predicted to give at the `r scales::dollar(10^tmpcutoff) %>% I()` level and `r tmpn <- error_compare %>% filter(pred.giving.bin >= tmpcutoff & log.giving.bin >= tmpcutoff) %>% nrow(); tmpn %>% I()` actually did, for a hit rate of `r (tmpn/tmpd) %>% scales::percent() %>% I()`.
  * Average giving in this group is `r ({error_compare %>% filter(pred.giving.bin >= tmpcutoff) %>% select(giving) %>% sum()}/tmpd) %>% scales::dollar() %>% I()` per person.
  * This method elimintates `r tmpd <- error_compare %>% filter(pred.giving.bin < tmpcutoff) %>% select(giving) %>% nrow(); tmpd %>% I()` entities, who gave a total of `r tmpn <- error_compare %>% filter(pred.giving.bin < tmpcutoff) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or an average of `r (tmpn/tmpd) %>% scales::dollar() %>% I()` per person.

```{r, fig.width = 6, fig.height = 8}
ev_bin_18 <- error_compare_18 %>%
  group_by(log.giving.bin, ev.bin) %>%
  summarise(n = length(gave), g = sum(gave))

ev_bin <- error_compare %>%
  group_by(log.giving.bin, ev.bin) %>%
  summarise(n = length(gave), g = sum(gave))

mingift <- -1
maxgift <- 8
grid.arrange(
ev_bin_18 %>%
  ggplot(aes(x = log.giving.bin, y = ev.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = 0:maxgift, minor_breaks = NULL, limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Actual giving', y = 'coef1 Expected value')
, ev_bin %>%
  ggplot(aes(x = log.giving.bin, y = ev.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = 0:maxgift, minor_breaks = NULL, limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Actual giving', y = 'coef2 Expected value')
)
```

__coef1__

  * Total giving during the out-of-sample time period was again `r tmpd <- error_compare_18%>% select(giving) %>% sum(); tmpd %>% scales::dollar() %>% I()`.
  * Entities with a score of `r tmpc2 <- 1; tmpc2 %>% I()` or higher (`r error_compare_18%>% filter(ev.bin >= tmpc2) %>% nrow()` total) gave `r tmpn <- error_compare_18%>% filter(ev.bin >= tmpc2) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or `r (tmpn/tmpd) %>% scales::percent() %>% I()` of the total.
  * Of these `r tmpd <- error_compare_18%>% filter(ev.bin >= tmpc2) %>% nrow(); tmpd` entities, `r tmpn <- error_compare_18%>% filter(ev.bin >= tmpc2 & giving > 0) %>% nrow(); tmpn` were donors (`r (tmpn/tmpd) %>% scales::percent()`), and `r tmpn <- error_compare_18%>% filter(ev.bin >= tmpc2 & giving >= 10^tmpcutoff) %>% nrow(); tmpn` gave at least `r scales::dollar(10^tmpcutoff)` (`r (tmpn/tmpd) %>% scales::percent()`).
  * Average giving in this group is `r ({error_compare_18%>% filter(ev.bin >= tmpc2) %>% select(giving) %>% sum()}/{error_compare_18%>% filter(ev.bin >= tmpc2) %>% select(giving) %>% nrow()}) %>% scales::dollar() %>% I()` per person.
  * This method elimintates `r tmpd <- error_compare_18%>% filter(ev.bin < tmpc2) %>% select(giving) %>% nrow(); tmpd %>% I()` entities, who gave a total of `r tmpn <- error_compare_18%>% filter(ev.bin < tmpc2) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or an average of `r (tmpn/tmpd) %>% scales::dollar() %>% I()` per person.

__coef2__

  * Total giving during the out-of-sample time period was again `r tmpd <- error_compare %>% select(giving) %>% sum(); tmpd %>% scales::dollar() %>% I()`.
  * Entities with a score of `r tmpc2 <- 1; tmpc2 %>% I()` or higher (`r error_compare %>% filter(ev.bin >= tmpc2) %>% nrow()` total) gave `r tmpn <- error_compare %>% filter(ev.bin >= tmpc2) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or `r (tmpn/tmpd) %>% scales::percent() %>% I()` of the total.
  * Of these `r tmpd <- error_compare %>% filter(ev.bin >= tmpc2) %>% nrow(); tmpd` entities, `r tmpn <- error_compare %>% filter(ev.bin >= tmpc2 & giving > 0) %>% nrow(); tmpn` were donors (`r (tmpn/tmpd) %>% scales::percent()`), and `r tmpn <- error_compare %>% filter(ev.bin >= tmpc2 & giving >= 10^tmpcutoff) %>% nrow(); tmpn` gave at least `r scales::dollar(10^tmpcutoff)` (`r (tmpn/tmpd) %>% scales::percent()`).
  * Average giving in this group is `r ({error_compare %>% filter(ev.bin >= tmpc2) %>% select(giving) %>% sum()}/{error_compare %>% filter(ev.bin >= tmpc2) %>% select(giving) %>% nrow()}) %>% scales::dollar() %>% I()` per person.
  * This method elimintates `r tmpd <- error_compare %>% filter(ev.bin < tmpc2) %>% select(giving) %>% nrow(); tmpd %>% I()` entities, who gave a total of `r tmpn <- error_compare %>% filter(ev.bin < tmpc2) %>% select(giving) %>% sum(); tmpn %>% scales::dollar() %>% I()`, or an average of `r (tmpn/tmpd) %>% scales::dollar() %>% I()` per person.

```{r, fig.width = 6, fig.height = 8}
both_bin_2018 <- error_compare_18%>%
  group_by(pred.giving.bin, ev.bin) %>%
  summarise(n = length(gave), g = sum(gave))

both_bin <- error_compare %>%
  group_by(pred.giving.bin, ev.bin) %>%
  summarise(n = length(gave), g = sum(gave))

mingift <- -1
maxgift <- 8

grid.arrange(
both_bin_2018 %>%
  ggplot(aes(x = pred.giving.bin, y = ev.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = 0:maxgift, minor_breaks = NULL, limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Predicted giving', y = 'coef1 Expected value')
, both_bin %>%
  ggplot(aes(x = pred.giving.bin, y = ev.bin, fill = n)) +
  geom_abline(color = 'purple', size = 2, alpha = .5) +
  geom_tile(alpha = .75) +
  geom_text(aes(label = n), color = 'white', size = 3) +
  scale_fill_gradient(name = 'count', trans = 'log10') +
  scale_x_continuous(breaks = mingift:maxgift, minor_breaks = NULL
                     , labels = 10^(mingift:maxgift) %>% scales::dollar(), limits = c(mingift, maxgift)) +
  scale_y_continuous(breaks = 0:maxgift, minor_breaks = NULL, limits = c(mingift, maxgift)) +
  coord_equal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = .5), axis.title.y = element_text(angle = 0, vjust = .5)) +
  labs(x = 'Predicted giving', y = 'coef2 Expected value')
)
```

__coef1__

 * Of the `r tmpsubset <- error_compare_18%>% filter(pred.giving.bin >= tmpcutoff & ev.bin < tmpc2); tmpsubset %>% nrow()` individuals with a relatively high predicted giving amount (`r scales::dollar(10^tmpcutoff)`) but a 0 expected value score, only `r tmpn <- tmpsubset %>% filter(giving > 0) %>% nrow(); tmpn` were actually donors (`r (tmpn/nrow(tmpsubset)) %>% scales::percent()`), giving a total of `r tmpsubset %>% filter(giving > 0) %>% select(giving) %>% sum() %>% scales::dollar()`.

__coef2__

 * Of the `r tmpsubset <- error_compare %>% filter(pred.giving.bin >= tmpcutoff & ev.bin < tmpc2); tmpsubset %>% nrow()` individuals with a relatively high predicted giving amount (`r scales::dollar(10^tmpcutoff)`) but a 0 expected value score, only `r tmpn <- tmpsubset %>% filter(giving > 0) %>% nrow(); tmpn` were actually donors (`r (tmpn/nrow(tmpsubset)) %>% scales::percent()`), giving a total of `r tmpsubset %>% filter(giving > 0) %>% select(giving) %>% sum() %>% scales::dollar()`.

This compares favorably with [last year's results](https://phively.github.io/ksm-models/pg-cultivation-score-fy18/06%20KSM%20model%20tuning.nb.html#combined_model).

Conclusion: using the coef1 model results in "missing" fewer donors.

# Score correlations

```{r}
# EOFY18 scores
fy18_ev_export <- read.csv(file = 'data/2019-01-09 export.csv')

yoy_comparison <- fy18_ev_export %>%
  select(ID_NUMBER, ev) %>%
  inner_join(
    output_data %>% select(ID_NUMBER, ev.coef1) %>% mutate(ID_NUMBER = as.numeric(ID_NUMBER))
    , by = 'ID_NUMBER'
  )  
```
```{r}
# EOFY18 versus EOFY19 pairwise scores by donor
yoy_comparison %>%
  ggplot(aes(x = ev, y = ev.coef1)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(title = 'EOFY18 versus EOFY19 pairwise scores by donor', x = 'FY18', y = 'FY19')
```
```{r}
lm(ev.coef1 ~ ev, data = yoy_comparison) %>% summary()
```

That's a very high correlation between PFY and CFY scores, as expected.

# Save results

```{r}
# Merged data frame
output_data <- data.frame(
  est.giving.coef1 = fy18_amt
  , est.p.coef1 = fy18_prb
) %>% mutate(
  ev.coef1 = est.giving.coef1 * est.p.coef1
  , est.giving.bin.coef1 = trunc(est.giving.coef1)
  , est.p.bin.coef1 = trunc(est.p.coef1 * 10)
  , ev.bin.coef1 = trunc(ev.coef1)
  , est.giving.coef2 = preds_amt
  , est.p.coef2 = preds_prb
  , ev.coef2 = est.giving.coef2 * est.p.coef2
  , est.giving.bin.coef2 = trunc(est.giving.coef2)
  , est.p.bin.coef2 = trunc(est.p.coef2 * 10)
  , ev.bin.coef2 = trunc(ev.coef2)
) %>% cbind(
  new.modeling.data
)
```

```{r}
# Master data file
output_data %>% write.csv(
  file = 'data/2019-10-17 predictions.csv'
  , row.names = FALSE
)
```

```{r}
# IDs and scores only
full.data %>% filter(
  # Exclude deceased
  RECORD_STATUS_CODE != 'D'
) %>% select(
  HOUSEHOLD_ID
  , ID_NUMBER
) %>% mutate(
  HOUSEHOLD_ID = HOUSEHOLD_ID %>% as.numeric()
) %>% left_join(
  output_data %>% mutate(
      HOUSEHOLD_ID = ID_NUMBER %>% as.numeric()
    ) %>%
    select(
      HOUSEHOLD_ID
      , est.giving.coef1
      , est.p.coef1
      , ev.coef1
      , est.giving.bin.coef1
      , est.p.bin.coef1
      , ev.bin.coef1
    ) 
  , by = c('HOUSEHOLD_ID')
) %>% write.csv(
  file = 'data/2019-10-17 export.csv'
  , row.names = FALSE
)
```
